% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_homogeneous_md_c1_c2_c3.R
\name{assumptions.wei_series_homogeneous_md_c1_c2_c3}
\alias{assumptions.wei_series_homogeneous_md_c1_c2_c3}
\title{Assumptions for \code{wei_series_homogeneous_md_c1_c2_c3} model.}
\usage{
\method{assumptions}{wei_series_homogeneous_md_c1_c2_c3}(model, ...)
}
\arguments{
\item{model}{the likelihood model object}

\item{...}{additional arguments (ignored)}
}
\value{
character vector of assumptions
}
\description{
Returns the assumptions made by this likelihood model.
}
