% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{cdf.series}
\alias{cdf.series}
\title{Cumulative distribution function (cdf) for a general series system.}
\usage{
\method{cdf}{series}(x)

\method{cdf}{series}(x)
}
\arguments{
\item{x}{series object}

\item{object}{series object}
}
\value{
cdf function, a function of \code{t} of times, parameter vector
\code{theta}, and logical \code{log} indicating whether to return the log of
the cdf
}
\description{
Cumulative distribution function (cdf) for a general series system.

Cumulative distribution function (cdf) for \code{series} systems.
}
