% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle.R
\name{confint.md_mle}
\alias{confint.md_mle}
\title{Method to obtain the confidence intervals of the parameter values of a
masked data estimator, \code{md_estimate}.}
\usage{
\method{confint}{md_mle}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The \code{md_estimate} object to compute the confidence intervals for}

\item{parm}{Unused}

\item{level}{Confidence level, defaults to 0.95 (alpha=.05)}

\item{...}{Additional arguments to pass.}
}
\description{
Method to obtain the confidence intervals of the parameter values of a
masked data estimator, \code{md_estimate}.
}
