% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{density.series}
\alias{density.series}
\title{Probability density function (pdf) for a general series system.}
\usage{
\method{density}{series}(object)
}
\arguments{
\item{object}{series object}
}
\value{
pdf function
}
\description{
Probability density function (pdf) for a general series system.
}
