% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{dexp_series}
\alias{dexp_series}
\title{pdf for exponential series.}
\usage{
dexp_series(t, rates, log = F)
}
\arguments{
\item{t}{series system lifetime}

\item{rates}{rate parameters for exponential component lifetimes}

\item{log}{return the log of the pdf}
}
\description{
pdf for exponential series.
}
