% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{dgeneral_series}
\alias{dgeneral_series}
\title{pdf for general series}
\usage{
dgeneral_series(t, theta, nparams, h, R)
}
\arguments{
\item{t}{series system lifetime}

\item{theta}{parameter vector}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{h}{list of hazard functions}

\item{R}{list of reliability functions}
}
\description{
pdf for general series
}
