% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series.R
\name{dweibull_series}
\alias{dweibull_series}
\title{pdf for weibull series}
\usage{
dweibull_series(t, scales, shapes)
}
\arguments{
\item{t}{series system lifetime}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
pdf for weibull series
}
