% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_data_3.R
\docType{data}
\name{exp_series_data_3}
\alias{exp_series_data_3}
\title{Masked data for a 10-out-of-10 (series system) with exponentially distributed nodes.}
\format{
A data frame with 100000 rows and 23 variables:
\describe{
\item{s}{Real observable variable, system lifetime}
\item{k}{Integer latent variable, the failed node}
\item{w}{Integer observable variable, number of candidates}
\item{t.1}{Real latent variable, lifetime of the 10 nodes}
\item{t.2}{Real latent variable, lifetimes of the 10 nodes}
\item{t.3}{Real latent variable, lifetimes of the 10 nodes}
\item{t.4}{Real latent variable, lifetimes of the 10 nodes}
\item{t.5}{Real latent variable, lifetimes of the 10 nodes}
\item{t.6}{Real latent variable, lifetimes of the 10 nodes}
\item{t.7}{Real latent variable, lifetimes of the 10 nodes}
\item{t.8}{Real latent variable, lifetimes of the 10 nodes}
\item{t.9}{Real latent variable, lifetimes of the 10 nodes}
\item{t.10}{Real latent variable, lifetimes of the 10 nodes}
\item{c.1}{Boolean observable variable}
\item{c.2}{Boolean observable variable}
\item{c.3}{Boolean observable variable}
\item{c.4}{Boolean observable variable}
\item{c.5}{Boolean observable variable}
\item{c.6}{Boolean observable variable}
\item{c.7}{Boolean observable variable}
\item{c.8}{Boolean observable variable}
\item{c.9}{Boolean observable variable}
\item{c.10}{Boolean observable variable}
}
}
\source{
\url{https://github.com/queelius/masked.data/blob/master/data-raw/exp_series_data_3_gen.R}
}
\usage{
exp_series_data_3
}
\description{
Masked data containing a sample of 100000 system lifetimes and other attributes
where the system is parameterized by \code{theta=c(3,5,4,6,7,2,8,9,10,11)} and
candidate model is m0.
}
\details{
Candidate set sizes are randomly drawn from {2,3,4,5,6,7,8,9}.
\code{t.j} is the lifetime of node \code{j}. If \code{c.j} is \code{TRUE},
that indicates node \code{j} is in candidate set.
}
\keyword{datasets}
