% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_data_5.R
\docType{data}
\name{exp_series_data_5}
\alias{exp_series_data_5}
\title{Masked data for a 3-out-of-3 (series) system with exponentially distributed
nodes.}
\format{
A data frame with 100000 rows and 13 variables:
\describe{
\item{s}{Real observable variable, system lifetime}
\item{k}{Integer latent variable, the failed node}
\item{w}{Integer observable variable, number of candidates}
\item{t.1}{Real latent variable, lifetime of node 1}
\item{t.2}{Real latent variable, lifetime of node 2}
\item{t.3}{Real latent variable, lifetime of node 3}
\item{c.1}{Boolean observable variable, TRUE indicates node 1 is in candidate set}
\item{c.2}{Boolean observable variable, TRUE indicates node 2 is in candidate set}
\item{c.3}{Boolean observable variable, TRUE indicates node 3 is in candidate set}
}
}
\source{
\url{https://github.com/queelius/masked.data/blob/master/data-raw/exp_series_data_5_gen.R}
}
\usage{
exp_series_data_5
}
\description{
Masked data containing the system lifetime and other attributes of 250
series parameterized by \code{theta=c(2,4,8)} and candidate model m0.
Candidate set sizes were randomly drawn from \code{{1,2,3}}. The seed for
the PRNG is 1233445.
}
\keyword{datasets}
