% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system_md_C1_C2_C3.R
\name{exp_series_system_md_C1_C2_C3_mss}
\alias{exp_series_system_md_C1_C2_C3_mss}
\title{Generates a set of minimally sufficient statistic for an exponential series
system with respect to rate parameter for masked data with candidate sets
that satisfy conditions C1, C2, and C3.}
\usage{
exp_series_system_md_C1_C2_C3_mss(df, sys.var = "t", cand.var = "x")
}
\arguments{
\item{df}{data frame}

\item{sys.var}{system lifetime column name, defaults to "t"}

\item{cand.var}{prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.}
}
\value{
sufficient statistics
}
\description{
Generates a set of minimally sufficient statistic for an exponential series
system with respect to rate parameter for masked data with candidate sets
that satisfy conditions C1, C2, and C3.
}
