% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectation}
\alias{expectation}
\title{Expectation operator with respect to pdf \code{f} of function \code{g},
i.e., \code{E(g(T))}, assuming \code{T ~ f(t)}.}
\usage{
expectation(f, g = function(t) t, lower = 0, upper = Inf)
}
\arguments{
\item{f}{pdf}

\item{g}{characteristic function of interest, defaults to T}

\item{lower}{lower limit of support of \code{f}, defaults to 0}

\item{upper}{upper limit of support of \code{f}, defaults to infinity}
}
\description{
Expectation operator with respect to pdf \code{f} of function \code{g},
i.e., \code{E(g(T))}, assuming \code{T ~ f(t)}.
}
