% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series_system.R
\name{Exponential series MLE}
\alias{Exponential series MLE}
\title{Exponential series MLE}
\description{
Maximum likelihood estimation functions for
Exponential series systems from masked data. Functions
include the log-likelihood, score, and FIM functions.
}
\details{
Masked component data approximately satisfies the
following conditions:

C1: Pr(K in C) = 1
C2: Pr(C=c | K=j, T=t) = Pr(C=c | K=j', T=t)
for any j, j' in c.
C3: masking probabilities are independent of theta
}
\seealso{
\code{\link{md_loglike_exp_series_C1_C2_C3}},
\code{\link{md_score_exp_series_C1_C2_C3}},
\code{\link{md_mle_exp_series_C1_C2_C3}}
\code{\link{md_fim_exp_series_C1_C2_C3}}
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{exponential,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
