% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{Exponential series system}
\alias{Exponential series system}
\title{Exponential series systems}
\description{
This file contains functions related to the Exponential series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions for the Exponential series distribution.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{exponential,}
\keyword{series,}
\keyword{statistics}
