% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system.R
\name{Exponential series}
\alias{Exponential series}
\title{Exponential series systems}
\description{
This file contains functions related to the Exponential series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions for the Exponential series distribution.
}
\details{
For parameter estimation from masked data, see the file
\code{md_exp_series_system.R}.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{exponential,}
\keyword{series,}
\keyword{statistics}
