% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{fisher_info.exp_dist}
\alias{fisher_info.exp_dist}
\title{Method to obtain the fisher information of the \code{x} object that
represents an \code{exp_dist} object.}
\usage{
\method{fisher_info}{exp_dist}(x, ...)
}
\description{
Method to obtain the fisher information of the \code{x} object that
represents an \code{exp_dist} object.
}
