% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{fisher_info.md_estimate}
\alias{fisher_info.md_estimate}
\title{Method to obtain the fisher infromation matrix of an md_estimate.}
\usage{
\method{fisher_info}{md_estimate}(estimate, ...)
}
\arguments{
\item{estimate}{The object to obtain the fisher information of}
}
\description{
Method to obtain the fisher infromation matrix of an md_estimate.
}
