% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle.R
\name{fisher_info.md_mle}
\alias{fisher_info.md_mle}
\title{Method to obtain the fisher information matrix of an \code{md_mle}.}
\usage{
\method{fisher_info}{md_mle}(x, ...)
}
\arguments{
\item{x}{The \code{md_mle} object to obtain the fisher information of}

\item{...}{Additional arguments to pass.}
}
\description{
Method to obtain the fisher information matrix of an \code{md_mle}.
}
