% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{fisher_info}
\alias{fisher_info}
\title{Generic method for obtaining the fisher information
matrix of an estimator, if supported.}
\usage{
fisher_info(x, ...)
}
\arguments{
\item{x}{The object to obtain the fisher information of}
}
\description{
Generic method for obtaining the fisher information
matrix of an estimator, if supported.
}
