% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_general_series.R
\name{General series MLE}
\alias{General series MLE}
\title{Maximum likelihood estimation functions for general series systems
from masked data.}
\description{
Functions include the log-likelihood, score, and MLE functions.
}
\seealso{
\code{\link{md_loglike_general_series_C1_C2_C3}},
\code{\link{md_score_general_series_C1_C2_C3}},
\code{\link{md_mle_general_series_C1_C2_C3}}
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
