% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{General series system}
\alias{General series system}
\title{General series system}
\description{
This file contains functions related to a general series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions. It models the concept of an Algebraic Distribution, in that
implements some generic methods in \code{algebraic.dist}.
}
\details{
It is based strictly on the paper "Series system estimation from masked
data" by Alex Towell. The paper is available at \url{?}.

For a likelihood model (for parameter estimation) from masked data
(competing risks) from candidate sets satisfying conditions C1, C2,
and C3, see the file \code{series_md_c1_c2_c3.R}.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{series,}
\keyword{statistics}
