% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{General series}
\alias{General series}
\title{General series}
\description{
This file contains functions related to a general series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions.
}
\details{
For parameter estimation from masked data, see the file \code{md_general_series_mle.R}.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{series,}
\keyword{statistics}
