% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guo_weibull_series_md}
\alias{guo_weibull_series_md}
\title{Real-world masked data. The source is from the paper,
"Estimating Component Reliabilities from Incomplete System Failure Data",
Table 2 (Example Data for a Series System).}
\format{
A data frame.
\describe{
\item{t}{Real unobservable variable, system lifetime}
\item{x1}{Boolean observable variable, TRUE indicates component 1 is in candidate set}
\item{x2}{Boolean observable variable, TRUE indicates component 2 is in candidate set}
\item{x3}{Boolean observable variable, TRUE indicates component 3 is in candidate set}
}
}
\source{
\url{https://github.com/queelius/series_system_estimation_masked_data/blob/master/data-raw/guo_weibull_series_md.R}
}
\usage{
guo_weibull_series_md
}
\description{
They assume the data is compatible with a series system of 3 components
w/Weibull distributed lifetimes and candidate sets that satisfy conditions
C1, C2, and C3.
}
\keyword{datasets}
