% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{hazard.exp_dist}
\alias{hazard.exp_dist}
\title{Method to obtain the hazard function of
an \code{exp_dist} object.}
\usage{
\method{hazard}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The \code{exp_dist} object to obtain the hazard function of}
}
\description{
Method to obtain the hazard function of
an \code{exp_dist} object.
}
