% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{hazard.exp_series_dist}
\alias{hazard.exp_series_dist}
\title{Method to obtain the hazard function of
an \code{exp_series_dist} object.}
\usage{
\method{hazard}{exp_series_dist}(x, ...)
}
\arguments{
\item{x}{The \code{exp_series_dist} object to obtain the hazard function of}
}
\description{
Method to obtain the hazard function of
an \code{exp_series_dist} object.
}
