% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{hazard}
\alias{hazard}
\title{Generic method for obtaining the hazard function of
a random variable.}
\usage{
hazard(x, ...)
}
\arguments{
\item{x}{The object to obtain the hazard function of}
}
\description{
Generic method for obtaining the hazard function of
a random variable.
}
