% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{hazard.series}
\alias{hazard.series}
\title{Hazard function for a general series system.}
\usage{
\method{hazard}{series}(object)

\method{hazard}{series}(object)
}
\arguments{
\item{object}{series object}

\item{t}{series system lifetime}

\item{theta}{parameter vector}
}
\value{
hazard function for series system
}
\description{
Hazard function for a general series system.

Hazard function for \code{series} system
}
