% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{hazard_exp_series}
\alias{hazard_exp_series}
\title{Hazard function for exponential series.}
\usage{
hazard_exp_series(t, rates, log.p = F)
}
\arguments{
\item{t}{series system lifetime}

\item{rates}{rate parameters for exponential component lifetimes}

\item{log.p}{return the log of the hazard function}
}
\description{
Hazard function for exponential series.
}
