% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system.R
\name{hazard_exp_series_system}
\alias{hazard_exp_series_system}
\title{Hazard function for exponential series.}
\usage{
hazard_exp_series_system(t, rates, log.p = FALSE)
}
\arguments{
\item{t}{Vector of series system lifetimes.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}

\item{log.p}{Logical; if TRUE, return the log of the hazard function.}
}
\value{
The hazard function evaluated at the specified lifetimes.
}
\description{
Hazard function for exponential series.
}
