% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{hazard_gen_series}
\alias{hazard_gen_series}
\title{Hazard function for general series}
\usage{
hazard_gen_series(t, theta, nparams, h)
}
\arguments{
\item{t}{series system lifetime}

\item{theta}{parameter vector}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparams[j]} parameters.}

\item{h}{list of hazard functions}
}
\description{
Hazard function for general series
}
