% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series_system.R
\name{hazard_weibull_series}
\alias{hazard_weibull_series}
\title{Hazard function for weibull series.}
\usage{
hazard_weibull_series(t, scales, shapes)
}
\arguments{
\item{t}{series system lifetime}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
Hazard function for weibull series.
}
