% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{hazards.series}
\alias{hazards.series}
\title{Component hazard function for a general series system.}
\usage{
hazards.series(x)
}
\arguments{
\item{x}{series object}
}
\value{
list of component hazard functions for series system
}
\description{
Component hazard function for a general series system.
}
