% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system_md_C1_C2_C3.R
\name{hess_loglik.exp_series_system_md_C1_C2_C3}
\alias{hess_loglik.exp_series_system_md_C1_C2_C3}
\title{Generates the observed information matrix (FIM) for an exponential series
system with respect to parameter \code{theta} for masked data with candidate
sets that approximately satisfy conditions C1, C2, and C3 with exact
failure times, or right-censoring (no component failure).}
\usage{
hess_loglik.exp_series_system_md_C1_C2_C3(md, options = list(), ...)
}
\arguments{
\item{md}{masked data with candidate sets that meet the
regular candidate model}
}
\value{
observed information matrix of type \verb{R^m -> R^(m x m)}
}
\description{
Generates the observed information matrix (FIM) for an exponential series
system with respect to parameter \code{theta} for masked data with candidate
sets that approximately satisfy conditions C1, C2, and C3 with exact
failure times, or right-censoring (no component failure).
}
