% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_candidate_set_utils.R
\name{informative_masking_by_rank}
\alias{informative_masking_by_rank}
\title{informative_masking_by_rank}
\usage{
informative_masking_by_rank(ts, alpha, beta)
}
\arguments{
\item{ts}{component failure times for the series system}

\item{alpha}{rate of change with respect to rank, as alpha -> 0 probabilities
go to discrete uniform distribution}

\item{beta}{max weight (for ranked 2 component)}
}
\description{
Returns \code{m=length(ts)} probabilities, where the j-th probability is the
probability that the j-th component will be in the candidate set.
probabilities are a function of rank (rather than by component failure
times).
}
\details{
The shape of the masking is given by two parametrs, \code{alpha}, which must be
non-negative, and \code{beta}, which must be between 0 and 1.
As \code{alpha} goes to 0, the probability vector approaches
(beta, beta, ..., 1, beta, ..., beta)
where the 1 is for the failed component and as \code{alpha} goes to infinity, it
assigns probability 1 and \code{beta} respectively to rank 1 (failed component)
and rank 2 components (and the remaining probabilities are zero).
}
