% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{inv_cdf.series}
\alias{inv_cdf.series}
\title{Quantile function (inverse of the cdf) for a general series system.
By definition, the quantile \verb{p * 100\%} is the value \code{t} that
satisfies \code{F(t) - p = 0}. We solve for \code{t} using Newton's method.}
\usage{
\method{inv_cdf}{series}(x, ...)
}
\arguments{
\item{...}{additional arguments to pass into \code{options}}

\item{p}{vector of probabilities.}

\item{theta}{parameter vector}

\item{object}{series object}
}
\description{
Quantile function (inverse of the cdf) for a general series system.
By definition, the quantile \verb{p * 100\%} is the value \code{t} that
satisfies \code{F(t) - p = 0}. We solve for \code{t} using Newton's method.
}
