% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{is_series}
\alias{is_series}
\title{Determine if an object is a \code{series} object.}
\usage{
is_series(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical, TRUE if \code{x} is a \code{series} object
}
\description{
Determine if an object is a \code{series} object.
}
