% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_candidate_set_utils.R
\name{kl_divergence_bernoulli}
\alias{kl_divergence_bernoulli}
\title{kl_divergence_bernoulli}
\usage{
kl_divergence_bernoulli(p, q)
}
\arguments{
\item{p}{probability of success for the first distribution}

\item{q}{probability of success for the second distribution}
}
\value{
KL divergence between the two distributions
}
\description{
KL divergence between two Bernoulli random vectors.
p = (p1,p2,...,pn) and q = (q1,q2,...,qn).
}
\examples{
kl_divergence_bernoulli(c(0.1,0.1), c(0.1,0.1)) = 0
}
