% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system_md_C1_C2_C3.R
\name{loglik.exp_series_system_md_C1_C2_C3}
\alias{loglik.exp_series_system_md_C1_C2_C3}
\title{Generates a log-likelihood function for an exponential series system with
respect to rate parameter for masked data with candidate sets that satisfy
conditions C1, C2, and C3.}
\usage{
loglik.exp_series_system_md_C1_C2_C3(model)
}
\arguments{
\item{md}{masked data}

\item{options}{list of options
\itemize{
\item \code{set.var} prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.
\item \code{sys.var} system lifetime (optionally right-censored) column name,
defaults to \code{t}
\item \code{delta.var} right-censoring indicator column name. If \code{NULL}, then
no right-censoring is assumed. If a system lifetime is
right-censored (\emph{not} observed), the right-censoring
indicator is \code{TRUE}, otherwise it is \code{FALSE}.
}}
}
\value{
log-likelihood function
}
\description{
Generates a log-likelihood function for an exponential series system with
respect to rate parameter for masked data with candidate sets that satisfy
conditions C1, C2, and C3.
}
