% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_md_c1_c2_c3.R
\name{loglik.series_md_c1_c2_c3}
\alias{loglik.series_md_c1_c2_c3}
\title{Masked data approximately satisfies the following set of conditions:
C1: Pr(K\link{i} in C\link{i}) = 1
C2: Pr(C\link{i}=c\link{i} | K\link{i}=j, T\link{i}=t\link{i}) =
Pr(C\link{i}=c\link{i} | K\link{i}=j', T\link{i}=t\link{i})
for any j,j' in c\link{i}.
C3: masking probabilities are independent of theta}
\usage{
loglik.series_md_c1_c2_c3(model)
}
\arguments{
\item{md}{masked data}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{h}{list of hazard functions}

\item{R}{list of survival functions, defaults to NULL, in which case we
generate the survival functions from \code{h}.}
}
\value{
a log-likelihood function with respect to theta
}
\description{
Masked data approximately satisfies the following set of conditions:
C1: Pr(K\link{i} in C\link{i}) = 1
C2: Pr(C\link{i}=c\link{i} | K\link{i}=j, T\link{i}=t\link{i}) =
Pr(C\link{i}=c\link{i} | K\link{i}=j', T\link{i}=t\link{i})
for any j,j' in c\link{i}.
C3: masking probabilities are independent of theta
}
