% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_md_c1_c2_c3.R
\name{loglik.wei_series_md_c1_c2_c3}
\alias{loglik.wei_series_md_c1_c2_c3}
\title{Log-likelihood method for \code{wei_series_md_c1_c2_c3} model.}
\usage{
\method{loglik}{wei_series_md_c1_c2_c3}(model, ...)
}
\arguments{
\item{model}{the likelihood model object}

\item{...}{additional arguments (passed to returned function)}
}
\value{
log-likelihood function that takes the following arguments:
\itemize{
\item \code{df}: masked data frame
\item \code{par}: parameter vector (shape1, scale1, shape2, scale2, ...)
\item \code{lifetime}: system lifetime column name (default from model)
\item \code{indicator}: right-censoring indicator column name (default from model)
\item \code{candset}: prefix of Boolean matrix encoding candidate sets
}
}
\description{
Returns a log-likelihood function for a Weibull series system with
respect to parameter vector (shape_1, scale_1, ..., shape_m, scale_m)
for masked data with candidate sets that satisfy conditions C1, C2, and C3.
}
