% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{loglike.series.exp}
\alias{loglike.series.exp}
\title{log-likelihood function of masked data
for a series system with exponentially distributed lifetimes.}
\usage{
loglike.series.exp(rate, masked.data)
}
\arguments{
\item{rate}{rate parameter}

\item{masked.data}{masked data}
}
\value{
log-likelihood of masked data wrt rate parameter
}
\description{
log-likelihood function of masked data
for a series system with exponentially distributed lifetimes.
}
