% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax_series_data_2.R
\docType{data}
\name{lomax_series_data_2}
\alias{lomax_series_data_2}
\title{Masked data for a series system with lomax distributed nodes and candidate
sets that model m0}
\format{
A data frame with 2000 rows and 9 variables:
\describe{
\item{s}{Real observable variable, system lifetime}
\item{k}{Integer latent variable, the failed node}
\item{w}{Integer observable variable, number of candidates}
\item{t.1}{Real latent variable, lifetime of node 1}
\item{t.2}{Real latent variable, lifetime of node 2}
\item{t.3}{Real latent variable, lifetime of node 3}
\item{c.1}{Boolean observable variable, TRUE indicates node 1 is in candidate set}
\item{c.2}{Boolean observable variable, TRUE indicates node 2 is in candidate set}
\item{c.3}{Boolean observable variable, TRUE indicates node 3 is in candidate set}
}
}
\source{
\url{https://github.com/queelius/masked.data/blob/master/data-raw/lomax_series_data_2_gen.R}
}
\usage{
lomax_series_data_2
}
\description{
Masked data containing the system lifetime and other attributes of 2000
series system with parameters \code{lambda=c(1,1.5,.75)} and \code{kappa=c(2,1.5,2.5)}. Every candidate
set (of model m0) has \code{w=2} candidate nodes.
}
\keyword{datasets}
