% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{make_exp_dist}
\alias{make_exp_dist}
\title{Construct exponential distribution object.}
\usage{
make_exp_dist(rate)
}
\arguments{
\item{rate}{failure rate}
}
\description{
Construct exponential distribution object.
}
