% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{make_exp_series}
\alias{make_exp_series}
\title{Construct exponential series object.}
\usage{
make_exp_series(rate)
}
\arguments{
\item{rate}{failure rates}
}
\description{
Construct exponential series object.
}
