% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{make_exp_series_dist}
\alias{make_exp_series_dist}
\title{Construct exponential series object.}
\usage{
make_exp_series_dist(rate)
}
\arguments{
\item{rate}{failure rates}
}
\description{
Construct exponential series object.
}
