% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_failure.R
\name{make_exp_series_node_failure_m0}
\alias{make_exp_series_node_failure_m0}
\title{Constructs a pdf object for the conditional node failure
in an exponential series system according to candidate model
m0, \code{f(k|c,s)}.}
\usage{
make_exp_series_node_failure_m0(exp_series)
}
\arguments{
\item{exp_series}{An \code{exp_series} object to construct the
node failure pdf object for.}
}
\description{
\code{f(k|s,c) = h_k(s)/h(s) I(k in c)}. This simplifies to
\code{f(k|c) = theta[k] / sum(theta[j],j in c)} for
the exponential series system.
}
