% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_failure.R
\name{make_exp_series_node_failure_m1}
\alias{make_exp_series_node_failure_m1}
\title{Constructs a pdf object for the conditional node failure
in an exponential series system according to candidate model
m1, \code{f(k|c,s,alpha)}.}
\usage{
make_exp_series_node_failure_m1(exp_series)
}
\description{
\code{f(k|s,c,alpha) = alpha h_k(s)/h(s) I(k in c) + (1-alpha) h_k(s)/h(s) I(k not in c)}.
This simplifies to \code{f(k|c,alpha) =
    alpha*theta[k]/sum(theta[j],j in c) I(k in c) +
    (1-alpha)*theta[k]/sum(theta[j],j in c) I(k not in c)}
for the exponential series system.
}
