% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{make_normal}
\alias{make_normal}
\title{Construct (multivariate or univariate) normal distribution object.}
\usage{
make_normal(mu, sigma = diag(length(mu)))
}
\arguments{
\item{theta}{parameter index}
}
\description{
Construct (multivariate or univariate) normal distribution object.
}
