% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_numeric_hessian}
\alias{make_numeric_hessian}
\title{Create numeric Hessian function from score function via Jacobian}
\usage{
make_numeric_hessian(score_fn, model)
}
\arguments{
\item{score_fn}{score function returned by score.* method}

\item{model}{likelihood model object (for extracting defaults)}
}
\value{
function(df, par, ...) that computes the Hessian matrix
}
\description{
Factory function for creating Hessian functions using numerical
differentiation of the score function. Used by Weibull models which compute
score analytically but need numerical Hessian.
}
\keyword{internal}
