% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked.data.R
\docType{package}
\name{masked.data}
\alias{masked.data}
\title{masked.data: A package for estimating parameters from masked data}
\description{
The \code{masked.data} package provides a general framework for
working with masked data and designing functions to solve
for the parameter's of latent node lifetime distributions
in a series system.
}
\section{data structures}{


\code{md_estimate}
\code{tbl_md}
}

\section{tbl_md tools}{


\code{md_write_csv}
}

\section{MLE point estimators}{


\code{md_mle_exp_series_m0}
\code{md_mle_exp_series_m1}
\code{md_mle_exp_series_m0}
\code{point.md_estimate}
}

\section{MLE interval estimators and covariance}{


\code{vcov.md_estimate}
\code{info.md_estimate}
\code{confint.md_estimate}
}

