% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mc_expectation}
\alias{mc_expectation}
\title{Monte carlo expectation of a random vector \code{T ~ sampler(1)} applied to a
function \code{g} where \code{g} should have a domain that is a superset of
the random vector's support. That is, we estimate \code{E(g(T))} with the
mean of \code{g(T[1]),...,g(T[n])}. We see that variance is
\code{1/n var(g(T[1]))}, which is inversely proportional to the sample size
\code{n}.}
\usage{
mc_expectation(sampler, g = function(t) t, n = 10000)
}
\arguments{
\item{sampler}{sampler, generates random vectors for a desired distribution}

\item{g}{characteristic function of interest, defaults to identity}

\item{n}{number of sample points, larger samples reduce variance}
}
\description{
Monte carlo expectation of a random vector \code{T ~ sampler(1)} applied to a
function \code{g} where \code{g} should have a domain that is a superset of
the random vector's support. That is, we estimate \code{E(g(T))} with the
mean of \code{g(T[1]),...,g(T[n])}. We see that variance is
\code{1/n var(g(T[1]))}, which is inversely proportional to the sample size
\code{n}.
}
