% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_models.R
\name{md_bernoulli_cand_C1_kld}
\alias{md_bernoulli_cand_C1_kld}
\title{md_bernoulli_cand_C1_kld}
\usage{
md_bernoulli_cand_C1_kld(
  md,
  p,
  d,
  eps = 1e-04,
  max_iter = 100000L,
  lr = 1,
  lambda = 1,
  alpha0 = 5,
  beta0 = 0.5,
  debug = F
)
}
\arguments{
\item{md}{component failure times for the series system}

\item{p}{numeric, defines \verb{P = (p, ..., p, 1, p, ..., p)}.}

\item{d}{numeric, the KL divergence from P = (p, p, ..., p, 1, p, ..., p)
to try to obtain}

\item{eps}{numeric, stopping condition.}

\item{max_iter}{Integer, maximum number of iterations before giving up.}

\item{lr}{numeric, learning rate.}

\item{lambda}{numeric, controls how much the two constraints are weighted.
Lower value specifies more enforcement of the KL-divergence
constraint being closer to \code{d}. Defaults to 1.}

\item{alpha0}{numeric, initial guess for \code{alpha} parameter of
\code{informative_masking_by_rank}.}

\item{beta0}{numeric, initial guess for \code{beta} parameter of
\code{informative_masking_by_rank}.}

\item{debug}{Logical, whether to output debugging information while running}
}
\description{
For each row (observation) in \code{md}, a probability \verb{Q = (q1, q2, ..., qm)} is
constructed such that \code{qj} is the probability that the j-th component is in
the candidate set, \code{qk = 1}, where \code{k} is failed component.
}
\details{
\code{Q} is an \emph{informed} candidate model that uses \code{informative_masking_by_rank}
to assign higher probabilities to components that failed earlier (which is
something we typically only know in, say, a simulation study).

The probabilities \code{Q} have two constraints on them. Let
\verb{P = (p, ..., p, 1, p, ..., p)} be the bernoulli candidate model that
satisfies conditions C1, C2, and C3. Then, the KL-divergence between \code{P}
and \code{Q} is as close as possible to \code{d} while satisfying \code{sum(P) == sum(Q)}.

For \code{d = 0}, \code{Q == P}. As \code{d} increases, \code{Q} becomes more informative about
the components. Given the structure of \code{informative_masking_by_rank}, it may
not be possible to satisfy every \code{d} specified, but we get as close as
we can, which should permit useful experiments.
}
