% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_decorators.R
\name{md_bernoulli_candidate_alpha_C1_C3}
\alias{md_bernoulli_candidate_alpha_C1_C3}
\title{Bernoulli candidate model that is a function of component
cause of failure.}
\usage{
md_bernoulli_candidate_alpha_C1_C3(md, gamma = NULL, alpha = NULL)
}
\arguments{
\item{md}{masked dataa data column 'k' for failed component and 'x'
for candidate sets}

\item{gamma}{probability vector}

\item{alpha}{probability vector, defaults to all 1's.}
}
\description{
Bernoulli candidate model that is a function of component
cause of failure \code{k} and parameters \code{alpha} and \code{gamma},
where if component cause of failure is indexed by \code{j}, then \code{j} is
in the candidate set with probability \code{alpha[j]}, and otherwise
the component indexed by \code{j} is in the candidate set with probability
\code{gamma[j]}.
}
\details{
If \code{all(alpha) = 1} then candidate sets satisfy C1.
If \code{all(gamma) = p} then candidiate sets satisfy C2.
Also, since we treat \code{gamma} as a numeric vector, it is not a function
of \code{theta} and therefore candidate sets satisfy C3.
}
