% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_decorators.R
\name{md_bernoulli_candidate_C1_C2_C3}
\alias{md_bernoulli_candidate_C1_C2_C3}
\title{Bernoulli candidate model that is a function of component
cause of failure.}
\usage{
md_bernoulli_candidate_C1_C2_C3(md, m, p = function(n) runif(n))
}
\arguments{
\item{md}{masked data, has a column \code{k} for the failed component index.}

\item{m}{integer, number of components in the candidate set.}

\item{p}{a function or procedure \code{p : integer -> [0,1]}, defaults to
\code{function(n) runif(n)}.}
}
\description{
Bernoulli candidate model that is a function of component
cause of failure \code{k} and parameters \code{m} and \code{p},
where if component cause of failure is indexed by \code{j}, then \code{j} is
in the candidate set and otherwise the component indexed by \code{j} is in
the candidate set with probability specified by \code{p}, which may either
by a procedure for sampling from a distribution with a codomain \code{[0,1]}
or a constant function.
}
