% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_decorators.R
\name{md_bernoulli_candidate_set_exp_series_generator}
\alias{md_bernoulli_candidate_set_exp_series_generator}
\title{Bernoulli candidate model that is a function of component
cause of failure and satisfies conditions \code{C1}, code{C2}, and \code{C3}.}
\usage{
md_bernoulli_candidate_set_exp_series_generator(
  k,
  m,
  alpha = 1,
  epsilon = 0.001
)
}
\arguments{
\item{k}{component cause of failure.}

\item{m}{number of components in the system}

\item{alpha}{defaults to \code{1}.}

\item{epsilon}{defaults to \code{1e-3}.}
}
\description{
Bernoulli candidate model that is a function of component
cause of failure \code{k} and parameters \code{alpha} and \code{epsilon},
where if component cause of failure is indexed by \code{j}, then \code{j} is
in the candidate set with probability \code{alpha[j]}.
}
\details{
\itemize{
\item If \code{alpha} is approximately \code{1}, then candidate sets
approximately satisfy \code{C1}.
\item If \code{epsilon} is sufficiently small, say \code{1e-3}, then candidate
sets approximately satisfy \code{C2}.
\item Candidate sets generated by this procedure always satisfy \code{C3}.
}
}
