% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_models.R
\name{md_block_candidate_m3}
\alias{md_block_candidate_m3}
\title{md_block_candidate_m3}
\usage{
md_block_candidate_m3(md, qvar = "q", compvar = "t")
}
\arguments{
\item{md}{masked data}

\item{qvar}{column prefix for component probabilities, defaults to \code{q},
e.g., \verb{q1, q2, ..., qm}.}

\item{compvar}{Prefix-code for the component lifetimes, defaults to \code{t},
e.g., \verb{t1, t2, ..., tm}.}
}
\description{
Block candidate model. It produces an MLE that is non-unique, i.e.,
if estimating exponential series system, the MLE
for λ₁, λ₂, and λ₃ satisfies
\verb{λ̂₁ + λ̂₂ = λ₁ + λ₂}
and
\verb{λ̂₃ = λ₃}.
}
\details{
This illustrates one of the conditions for the MLE to be unique:

\if{html}{\out{<div class="sourceCode">}}\preformatted{(1) multiple components must not always show together. if they
do, then the best we can learn is that they satisfy some hyper-plane
constraint.
}\if{html}{\out{</div>}}
}
