% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_set_models.R
\name{md_cand_sampler}
\alias{md_cand_sampler}
\title{md_cand_sampler}
\usage{
md_cand_sampler(md, qvar = "q", setvar = "x")
}
\arguments{
\item{md}{masked data.}

\item{qvar}{column prefix for component probabilities, defaults to \code{q},
e.g., \verb{q1, q2, ..., qm}.}

\item{setvar}{column prefix for candidate sets (as Boolean matrix), defaults
to \code{x}, e.g., \verb{x1, x2, ..., xm}.}
}
\description{
Candidate set generator. Requires columns for component probabilities
e.g., \verb{q1,...,qm} where \code{qj} is the probability that the jth component
will be in the corresponding candidate set generated for that observation
in the \code{md} table.
}
