% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_models.R
\name{md_candidate_m1}
\alias{md_candidate_m1}
\title{Decorates a masked data (input md) with candidate sets according to model m0.}
\usage{
md_candidate_m1(md)
}
\arguments{
\item{md}{masked data, a data frame object with column 'k' for failed component,
column 'w' for corresponding candidate set size, and column 'alpha' for
corresponding alpha probabilities}
}
\value{
alpha-masked data with candidate sets that model m1
}
\description{
Decorates a masked data (input md) with candidate sets according to model m0.
}
