% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_candidates_as_matrix}
\alias{md_candidates_as_matrix}
\title{Convert the columns corresponding to the
candidate matrix to a matrix object.}
\usage{
md_candidates_as_matrix(md)
}
\arguments{
\item{md}{masked data}
}
\value{
Candidate sets represented as a Boolean matrix
}
\description{
Convert the columns corresponding to the
candidate matrix to a matrix object.
}
