% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_candidates_to_strings}
\alias{md_candidates_to_strings}
\title{Candidate matrix to stringified vector of integers}
\usage{
md_candidates_to_strings(md)
}
\arguments{
\item{md}{masked data}
}
\description{
Candidate matrix to stringified vector of integers
}
