% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compat.R
\name{md_encode_matrix}
\alias{md_encode_matrix}
\title{Encode a matrix as a data frame with prefixed column names}
\usage{
md_encode_matrix(mat, var)
}
\arguments{
\item{mat}{matrix to encode}

\item{var}{character prefix for the column names}
}
\value{
a data frame with named columns
}
\description{
Converts a matrix to a data frame with columns named \verb{var1, var2, ...}.
}
