% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_exp_series}
\alias{md_exp_series}
\title{Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.}
\usage{
md_exp_series(n, theta, w, candidate_model = md_candidate_m0)
}
\arguments{
\item{n}{Integer. The sample size (each row is an observation).}

\item{theta}{Numeric vector. The j-th component has a failure rate theta_j.}

\item{w}{Integer vector. For the ith observation, generate w_j candidates.}

\item{candidate_model}{the candidate model, defaults to md_candidate_m0.}
}
\value{
masked data, a data frame of n observations, (s,k,t1,...,tm,c1,...,cm)
where k, t, and c are covariates (or predictors) of s,k,t1,...,tm.
}
\description{
Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.
}
\examples{
md <- md_exp_series(
    n=10,
    theta=c(1,2,3),
    w=rep(2,10),
    candidate_model=md_candidate_m1)
}
