% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_exp_series_component_failure_reg_cand}
\alias{md_exp_series_component_failure_reg_cand}
\title{Constructs a pdf (pmf) object for the conditional probability of component
failure \code{Pr{K\link{i}=j|C\link{i}=c\link{i},T\link{i}=t\link{i}) = h_j(t\link{i})/h(t\link{i}) I(j in c\link{i})}.
in an exponential series system for the regular candidate model.}
\usage{
md_exp_series_component_failure_reg_cand(theta)
}
\arguments{
\item{theta}{parameter value of \code{exp_series_dist}}
}
\description{
Constructs a pdf (pmf) object for the conditional probability of component
failure \code{Pr{K\link{i}=j|C\link{i}=c\link{i},T\link{i}=t\link{i}) = h_j(t\link{i})/h(t\link{i}) I(j in c\link{i})}.
in an exponential series system for the regular candidate model.
}
