% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_exp_series_decorator}
\alias{md_exp_series_decorator}
\title{Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.}
\usage{
md_exp_series_decorator(md, theta, metadata = T)
}
\arguments{
\item{md}{a masked data frame to decorate with exponential series data}

\item{theta}{component \code{j} has a failure rate \code{theta[j]}.}

\item{metadata}{Boolean. If \code{TRUE} writes meta-data for series system to
attributes of masked data.}

\item{cand_model}{function that accepts masked data as an argument.
the candidate model, defaults to \code{md_cand_m0}.
If set to \code{NULL}, then do not generate a candidate set.
\code{md_mle_exp_series} will treat such masked data as
a sample that includes every node as candidates.}
}
\value{
masked data populated with columns for \code{s}, \code{k}, and
a the matrix of component lifetimes encoded in the columns
\code{t.1,t.2,...,t.m}.
}
\description{
Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.
}
\examples{
md_exp_series_decorator(n=10,theta=c(1,2,3),w=rep(2,10))
}
