% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_exp_series_node_failure_m0}
\alias{md_exp_series_node_failure_m0}
\title{Constructs a pdf object for the conditional node failure
in an exponential series system according to candidate model
m0, \code{f(k|c,s)} = h_k(s)/h(s) I(k in c)}.}
\usage{
md_exp_series_node_failure_m0(theta)
}
\arguments{
\item{theta}{parameter value of \code{exp_series}}
}
\description{
This simplifies to \code{f(k|c) = theta[k] / sum(theta[j],j in c)} for
the exponential series system.
}
