% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_exp_series_system_failure_interval_m0}
\alias{md_exp_series_system_failure_interval_m0}
\title{Constructs the shortest interval for the system lifetime
given a candidate set under model m0 with a probability \code{p}
that the interval contains the system failure.}
\usage{
md_exp_series_system_failure_interval_m0(theta, p)
}
\arguments{
\item{theta}{parameter value of \code{exp_series}}

\item{p}{probability that system failure time is in the computed interval}
}
\description{
Constructs the shortest interval for the system lifetime
given a candidate set under model m0 with a probability \code{p}
that the interval contains the system failure.
}
