% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_fim_exp_series_C1_C2_C3}
\alias{md_fim_exp_series_C1_C2_C3}
\title{Generates the observed information matrix (FIM) for an exponential series system
with respect to parameter \code{theta} for masked data with candidate sets
that approximately satisfy conditions C1, C2, and C3.}
\usage{
md_fim_exp_series_C1_C2_C3(md, setvar = "x", sysvar = "t")
}
\arguments{
\item{md}{masked data with candidate sets that meet the
regular candidate model}

\item{setvar}{prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.}

\item{sysvar}{system lifetime (optionally right-censored) column name}
}
\value{
observed information matrix of type \verb{R^m -> R^(m x m)}
}
\description{
Generates the observed information matrix (FIM) for an exponential series system
with respect to parameter \code{theta} for masked data with candidate sets
that approximately satisfy conditions C1, C2, and C3.
}
