% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_fisher_scoring}
\alias{md_fisher_scoring}
\title{Fisher scoring algorithm.}
\usage{
md_fisher_scoring(theta0, info, score, eps = 1e-05)
}
\arguments{
\item{theta0}{initial guess of theta with p components}

\item{info}{information matrix function of type R^p -> R^(p x p)}

\item{score}{score function of type R^p -> R^p}

\item{eps}{stopping condition}
}
\value{
MLE estimate of theta
}
\description{
Fisher scoring algorithm.
}
