% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_tools.R
\name{md_fisher_scoring_loglike}
\alias{md_fisher_scoring_loglike}
\title{Fisher scoring algorithm using the log-likelihood (or kernel log-like)
function.}
\usage{
md_fisher_scoring_loglike(theta0, loglike, eps = 1e-05, gamma = 1)
}
\arguments{
\item{theta0}{initial guess of theta with p components}

\item{loglike}{information matrix function of type R^p -> R^(p x p)}

\item{eps}{stopping condition}
}
\value{
MLE estimate of theta
}
\description{
Fisher scoring algorithm using the log-likelihood (or kernel log-like)
function.
}
