% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{md_grad_mle_series_C1_C2_C3}
\alias{md_grad_mle_series_C1_C2_C3}
\title{Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.}
\usage{
md_grad_mle_series_C1_C2_C3(md, loglike, theta0, eps = 1e-05, max_iter = 0L)
}
\arguments{
\item{md}{masked data}

\item{loglike}{log-likelihood function}

\item{theta0}{initial guess for MLE}

\item{eps}{stopping condition, defaults to \code{1e-05}}

\item{max_iter}{stop if iterations reaches max_iter, default is 0 (infinity)}
}
\value{
numeric MLE estimate
}
\description{
Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.
}
