% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_info_exp_series_C1_C2_C3}
\alias{md_info_exp_series_C1_C2_C3}
\title{Generates the observed information matrix for an exponential series system
with respect to parameter \code{theta} for masked data with candidate sets
that approximately satisfy conditions C1, C2, and C3.}
\usage{
md_info_exp_series_C1_C2_C3(md)
}
\arguments{
\item{md}{masked data with candidate sets \code{x1,...,xm} that meet the
regular candidate model}
}
\value{
observed information matrix of type \code{R^m -> R^(m x m)}
}
\description{
Generates the observed information matrix for an exponential series system
with respect to parameter \code{theta} for masked data with candidate sets
that approximately satisfy conditions C1, C2, and C3.
}
