% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_info_exp_series_m0}
\alias{md_info_exp_series_m0}
\title{Expected information matrix for rate parameter
with respect to masked data of a series system
with exponentially distributed lifetimes and
candidate model m0.}
\usage{
md_info_exp_series_m0(md)
}
\arguments{
\item{md}{masked data for candidate model m0}
}
\value{
observed information matrix of type R^m -> R^(m x m)
}
\description{
Expected information matrix for rate parameter
with respect to masked data of a series system
with exponentially distributed lifetimes and
candidate model m0.
}
