% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_info_exp_series_reg_cand}
\alias{md_info_exp_series_reg_cand}
\title{Information matrix (observed) for rate parameter
with respect to masked data of a series system
with exponentially distributed lifetimes and
the regular candidate model.}
\usage{
md_info_exp_series_reg_cand(md)
}
\arguments{
\item{md}{masked data with candidate sets \code{x1,...,xm} that meet the
regular candidate model}
}
\value{
observed information matrix of type R^m -> R^(m x m)
}
\description{
Information matrix (observed) for rate parameter
with respect to masked data of a series system
with exponentially distributed lifetimes and
the regular candidate model.
}
