% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_is_masked_data}
\alias{md_is_masked_data}
\title{Test whether \code{x} is masked data}
\usage{
md_is_masked_data(x)
}
\arguments{
\item{x}{object to test}
}
\description{
An object is considered to be masked data if
it is a type of data frame (e.g., tibble)
and it has at least two columns for candidate
sets named \code{c.1} and \code{c.2} and a column for
system failure time named \code{s}.
}
