% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_kloglike_exp_series_m0_slow}
\alias{md_kloglike_exp_series_m0_slow}
\title{Kernel log-likelihood for masked data m0 for exponential series system.}
\usage{
md_kloglike_exp_series_m0_slow(md)
}
\arguments{
\item{md}{masked data for candidate model m0}
}
\description{
The log of the kernel of the likelihood function for masked data
for a series system with exponentially distributed lifetimes
and candidate sets that model m0.
}
\details{
This is the unoptimized version, which serves as a ground-truth
for testing a more efficient implementation.
}
