% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_lomax_series.R
\name{md_kloglike_lomax_series_m0_ref}
\alias{md_kloglike_lomax_series_m0_ref}
\title{Kernel log-likelihood for masked data m0 for lomax series system.}
\usage{
md_kloglike_lomax_series_m0_ref(md)
}
\arguments{
\item{md}{masked data for candidate model m0}
}
\description{
The log of the kernel of the likelihood function for masked data
for a series system with lomax distributed lifetimes
and candidate sets that model m0.
}
\details{
This is the unoptimized version, which serves as a ground-truth
for testing a more efficient implementation.
}
