% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_loglike_exp_series_C1_C2_C3}
\alias{md_loglike_exp_series_C1_C2_C3}
\title{Generates a log-likelihood for an exponential series system with respect to
parameter \code{theta} for masked data with candidate sets that approximately
satisfy conditions C1, C2, and C3.}
\usage{
md_loglike_exp_series_C1_C2_C3(md)
}
\arguments{
\item{md}{masked data}
}
\description{
Generates a log-likelihood for an exponential series system with respect to
parameter \code{theta} for masked data with candidate sets that approximately
satisfy conditions C1, C2, and C3.
}
