% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_loglike_exp_series_m0}
\alias{md_loglike_exp_series_m0}
\title{log-likelihood function of masked data for a series system
with exponentially distributed lifetimes.}
\usage{
md_loglike_exp_series_m0(md)
}
\arguments{
\item{md}{masked data for candidate model m0}
}
\value{
log-likelihood function of type R^m -> R
}
\description{
log-likelihood function of masked data for a series system
with exponentially distributed lifetimes.
}
