% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_loglike_exp_series_reg_cand}
\alias{md_loglike_exp_series_reg_cand}
\title{Log-likelihood for masked data with candidate sets obeying the
regular candidate set model for exponential series system.}
\usage{
md_loglike_exp_series_reg_cand(md)
}
\arguments{
\item{md}{masked data}
}
\description{
Log-likelihood for masked data with candidate sets obeying the
regular candidate set model for exponential series system.
}
