% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{md_loglike_general_series_C1_C2_C3}
\alias{md_loglike_general_series_C1_C2_C3}
\title{Masked data approximately satisfies the following set of conditions:
C1: \code{Pr(K[i] in C[i]) = 1}
C2: \code{Pr(C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])}
for any \code{j,j' in c[i]}.
C3: masking probabilities are independent of theta}
\usage{
md_loglike_general_series_C1_C2_C3(md, nparams, h, R = NULL)
}
\arguments{
\item{md}{masked data}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{h}{list of hazard functions}

\item{R}{list of survival functions, defaults to NULL, in which case we
generate the survival functions from \code{h}.}
}
\value{
a log-likelihood function with respect to theta given \code{md}
}
\description{
Masked data approximately satisfies the following set of conditions:
C1: \code{Pr(K[i] in C[i]) = 1}
C2: \code{Pr(C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])}
for any \code{j,j' in c[i]}.
C3: masking probabilities are independent of theta
}
