% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series.R
\name{md_loglike_weibull_series_C1_C2_C3}
\alias{md_loglike_weibull_series_C1_C2_C3}
\title{Masked data approximately satisfies the following set of conditions:
C1: \code{Pr(K[i] in C[i]) = 1}
C2: \code{Pr(C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])
    for any j, j' in c[i]}.
C3: masking probabilities are independent of theta}
\usage{
md_loglike_weibull_series_C1_C2_C3(md)
}
\arguments{
\item{md}{masked data}
}
\value{
a log-likelihood function with respect to theta given md
}
\description{
Masked data approximately satisfies the following set of conditions:
C1: \code{Pr(K[i] in C[i]) = 1}
C2: \code{Pr(C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])
    for any j, j' in c[i]}.
C3: masking probabilities are independent of theta
}
