% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_lomax_series.R
\name{md_lomax_series}
\alias{md_lomax_series}
\title{Generates masked data for a series system with lomax distributed
nodes and candidate sets according to candidate_model.}
\usage{
md_lomax_series(
  n,
  lambda,
  kappa,
  w,
  candidate_model = md_candidate_m0,
  metadata = T
)
}
\arguments{
\item{n}{Integer. The sample size (each row is an observation).}

\item{lambda}{Numeric vector.}

\item{kappa}{Numeric vector. The jth node is parameterized by theta_j := (lambda_j,kappa_j).}

\item{w}{Integer vector. For the ith observation, generate w_j candidates.}

\item{candidate_model}{Function that accepts masked data as an argument.
The candidate model, defaults to md_candidate_m0.
If set to NULL, then do not generate a candidate set.
md_mle_exp_series will treat such masked data as
a sample that includes every node as candidates.}

\item{metadata}{Boolean. If TRUE writes meta-data for series system to
attributes of masked data (tbl_md).}
}
\value{
masked data, a data frame of n observations, (s,k,t1,...,tm,c1,...,cm)
where k, t, and c are covariates (or predictors) of s,k,t1,...,tm.
}
\description{
Generates masked data for a series system with lomax distributed
nodes and candidate sets according to candidate_model.
}
\examples{
md_lomax_series(n=10,lambda=c(1,2,3),kappa=c(4,5,6),w=rep(2,10))
}
