% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_mle_exp_series_C1_C2_C3}
\alias{md_mle_exp_series_C1_C2_C3}
\title{Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.}
\usage{
md_mle_exp_series_C1_C2_C3(
  md,
  theta0,
  stop_cond = function(theta1, theta0) abs(max(theta1 - theta0)) < 1e-05,
  max_iter = 0L
)
}
\arguments{
\item{md}{masked data}

\item{theta0}{initial guess for MLE}

\item{stop_cond}{stopping condition, defaults to the absolute of the max component difference being less than 1e-05}

\item{max_iter}{stop if iterations reaches max_iterations.}
}
\value{
MLE estimate
}
\description{
Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.
}
