% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_mle_exp_series_m0}
\alias{md_mle_exp_series_m0}
\title{maximum likelihood estimator of the
parameters of a series system with
components that have exponentially
distributed lifetimes given a
sample of masked data.}
\usage{
md_mle_exp_series_m0(md, theta0 = NULL, eps = 1e-05)
}
\arguments{
\item{md}{masked data}

\item{theta0}{initial guess for MLE}

\item{eps}{stopping condition}
}
\value{
mle of theta given the observed masked data md
}
\description{
maximum likelihood estimator of the
parameters of a series system with
components that have exponentially
distributed lifetimes given a
sample of masked data.
}
