% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_mle_exp_series_reg_cand}
\alias{md_mle_exp_series_reg_cand}
\title{Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.}
\usage{
md_mle_exp_series_reg_cand(md, theta0, eps = 0.001, max_iter = 250L)
}
\arguments{
\item{md}{masked data}

\item{theta0}{initial guess for MLE}

\item{eps}{stopping condition}

\item{max_iterations}{stop if iterations reaches max_iterations.}
}
\value{
MLE estimate
}
\description{
Maximum likelihood estimator of the parameters of a series
system with nodes that have exponentially distributed
lifetimes given a sample of masked data according to
candidate model m0.
}
