% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{md_mle_series_C1_C2_C3}
\alias{md_mle_series_C1_C2_C3}
\title{Maximum likelihood estimator of the parameters of a series
system given a sample of masked data according to candidate sets
that are generated in a way that is compatible with conditions C1, C2, and
C3 so that we do not need to know or estimate the distribution of candidate
sets.}
\usage{
md_mle_series_C1_C2_C3(
  loglike,
  theta0,
  theta.restart = NULL,
  sup = function(theta) all(theta > 0),
  eps = 1e-05,
  restarts = 100L,
  max_iter = 250L,
  eta = 1,
  r = 0.5
)
}
\arguments{
\item{loglike}{log-likelihood function}

\item{theta0}{initial guess for MLE}

\item{theta.restart}{a function \code{() -> R^m} for generating new theta0 for restarts, defaults to randomly generating values}

\item{sup}{support function, defaults to all components of \code{theta} being positive non-zero}

\item{eps}{stopping condition, defaults to \code{1e-05}}

\item{restarts}{number of restarts to do, defaults to 100}

\item{max_iter}{stop if iterations reaches max_iter, default is 0 (infinity)}

\item{eta}{learning rate, defaults to 1}

\item{r}{backtracking line search parameter}
}
\value{
numeric MLE estimate
}
\description{
Maximum likelihood estimator of the parameters of a series
system given a sample of masked data according to candidate sets
that are generated in a way that is compatible with conditions C1, C2, and
C3 so that we do not need to know or estimate the distribution of candidate
sets.
}
