% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_node_times_as_matrix}
\alias{md_node_times_as_matrix}
\title{Convert the columns corresponding to the
node times matrix to a matrix object.}
\usage{
md_node_times_as_matrix(md)
}
\arguments{
\item{md}{masked data}
}
\value{
Node times represented as a real matrix
}
\description{
Convert the columns corresponding to the
node times matrix to a matrix object.
}
