% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{md_num_nodes}
\alias{md_num_nodes}
\title{Retrieve the number of nodes implicitly
defined by the masked data input 'md'.}
\usage{
md_num_nodes(md)
}
\arguments{
\item{md}{masked data}
}
\value{
number of nodes in the series system
}
\description{
Retrieve the number of nodes implicitly
defined by the masked data input 'md'.
}
