% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_observed_info_exp_series_m0}
\alias{md_observed_info_exp_series_m0}
\title{Observed information matrix of the rate parameter
of the series system with exponentially
distributed component lifetimes given
masked data with candidate sets according to model m0.}
\usage{
md_observed_info_exp_series_m0(md)
}
\arguments{
\item{md}{masked data}
}
\value{
observed info
}
\description{
Observed information matrix of the rate parameter
of the series system with exponentially
distributed component lifetimes given
masked data with candidate sets according to model m0.
}
