% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_read_json}
\alias{md_read_json}
\title{Read masked data from a JSON file.
If the JSON file has a 'dataset' field,
then each member of this field is assumed
to refer to a CSV file to read a masked
data sample from.}
\usage{
md_read_json(filename)
}
\arguments{
\item{filename}{filename for csv}
}
\value{
list of masked data objects
}
\description{
Any metadata in the JSON
file is inserted into the attributes
of the masked data samples.
}
