% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_models.R
\name{md_reg_cand}
\alias{md_reg_cand}
\title{Regular candidate model}
\usage{
md_reg_cand(md, m)
}
\arguments{
\item{md}{masked data, data frame object with column \code{k} for failed component}

\item{m}{number of components in the series system}
}
\value{
masked data
}
\description{
Decorates masked data object \code{md} with candidate sets according to
the regular candidate model.
}
