% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_masked_component_cause.R
\name{md_sample_candidates}
\alias{md_sample_candidates}
\title{Empirical sampler for C\link{i} | K\link{i} = k given a masked
data frame with component cause of failures shown.}
\usage{
md_sample_candidates(n, df, t, k, lifetime = "t", candset = "x", nbins = 10)

md_sample_candidates(n, df, t, k, lifetime = "t", candset = "x", nbins = 10)
}
\arguments{
\item{df}{data frame (sample) that we used to estimate C\link{i} | T\link{i} = t\link{i}, K\link{i} = j
\code{df} should only contain data in which the system failed, rather than being
right-censored.}

\item{t}{observed lifetime, defaults to NA (unknown or unconditional)}

\item{k}{component cause of failure, defaults to NA (unknown)}

\item{candset}{column prefix for candidate sets, defaults to \code{x},
e.g., \verb{x1, x2, x3}.}

\item{nbins}{number of bins to use for discretizing the component lifetimes}

\item{cause}{column name for component cause of failure, defaults to \code{k}}
}
\value{
a data frame of candidate sets sampled from the empirical
distribution of C\link{i} | K\link{i} = k
}
\description{
In the likelihood model, we assume rather that
C\link{i} | T\link{i} = t, K\link{i} = k satisfies certain assumptions.
For tractability and efficiency, we're discarding the conditional relation on
T\link{i} and only keeping the conditional relation on K\link{i} by using the
empirical distribution of C\link{i} | K\link{i} = k.

In the likelihood model, we assume that C\link{i} | T\link{i} = t, K\link{i} = k satisfies
certain assumptions. If component cause of failure is known, then we can
sample from the conditional distribution of C\link{i} | T\link{i} = t, K\link{i} = k,
otherwise we can sample from the empirical distribution of C\link{i} | T\link{i} = t\link{i},
which may still be reasonable, since by Condition 2,
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j} =
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j'}  for all j, j' in c\link{i}.
Of course, we may violate Condition 1, Pr{K\link{i} in C\link{i}} = 1.
}
\details{
If k is not known, then we can sample from the empirical distribution of
C\link{i}, which may still be reasonable, since by Condition 2,
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j} =
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j'}  for all j, j' in c\link{i},
and so in general varying the component cause among the components
in the candidate set does not change the probability of the candidate set.
Of course, we may violate Condition 1, Pr{K\link{i} in C\link{i}} = 1.

If the model is more informed, then dropping the conditional relation on
K\link{i} will lose information, and may bias the results in complex ways.

Note that if |c\link{i}| = 1, then we can sample from the empirical
distribution of C\link{i} | T\link{i} = t\link{i}, K\link{i} = j, since there is only
one possible value for K\link{i}, K\link{i} = j if c\link{i} = {j}. Otherwise,
if |c\link{i}| > 1, and we don't know the component cause of failure,
then we can sample from the empirical distribution of C\link{i} | T\link{i} = t\link{i}.

In the semi-parametric bootstrap, we generate the samples ourselves
from our estimate of theta, so we can sample from the conditional
distribution of C\link{i} | T\link{i} = t\link{i}, K\link{i} = j, since we know the
simulated component cause of failure. This is the main reason we
want to sample C\link{i} | T\link{i} = t\link{i}, K\link{i} = j, so this is fine.

NOTE: If the sample contains other kinds of likelihood contributions, like
for instance right-censoring, then please remove those rows before
calling this function, e.g., df <- df\link{df$delta, }. Otherwise, the
function will sample from the empirical distribution of C\link{i} conditioned on
other kinds of likelihood contributions, which is not what we want.
}
