% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_score_exp_series_C1_C2_C3}
\alias{md_score_exp_series_C1_C2_C3}
\title{Generates a score function for an exponential series system (or competing risks)
with respect to parameter \code{theta} for masked component failure (or masked
competing risks) with candidate sets (risks) that approximately satisfy conditions
C1, C2, and C3.}
\usage{
md_score_exp_series_C1_C2_C3(md)
}
\arguments{
\item{md}{right-censored failure-time data with masked competing risks}
}
\value{
score function of type \code{R^m -> R}
}
\description{
Generates a score function for an exponential series system (or competing risks)
with respect to parameter \code{theta} for masked component failure (or masked
competing risks) with candidate sets (risks) that approximately satisfy conditions
C1, C2, and C3.
}
