% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_score_exp_series_m0}
\alias{md_score_exp_series_m0}
\title{score function of masked data for a series system
with exponentially distributed lifetimes.}
\usage{
md_score_exp_series_m0(md)
}
\arguments{
\item{md}{masked data for candidate model m0}
}
\value{
score function of type R^m -> R
}
\description{
score function of masked data for a series system
with exponentially distributed lifetimes.
}
