% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_score_exp_series_reg_cand}
\alias{md_score_exp_series_reg_cand}
\title{score function of masked data for a series system
with exponentially distributed lifetimes.}
\usage{
md_score_exp_series_reg_cand(md)
}
\arguments{
\item{md}{masked data for regular candidate model}
}
\value{
score function of type R^m -> R
}
\description{
score function of masked data for a series system
with exponentially distributed lifetimes.
}
