% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_weibull_series.R
\name{md_score_weibull_series_C1_C2_C3}
\alias{md_score_weibull_series_C1_C2_C3}
\title{Score function generator (gradient of the log-likelihood) for Weibull series
system on masked data, where the masked data is in the form of right-censored
system lifetimes and masked component cause of failure. Note that if the right
censoring indicator variable \code{delta} is missing, we assume that all system
lifetimes are observed.}
\usage{
md_score_weibull_series_C1_C2_C3(md, sysvar = "t", setvar = "x")
}
\arguments{
\item{md}{masked data}

\item{sysvar}{name of the system lifetime column}

\item{setvar}{prefix symbol for the candidate sets column}
}
\value{
a score function with respect to theta given md
}
\description{
Masked component data \code{md} approximately satisfies the following conditions:
C1: Pr(K in C) = 1
C2: Pr(C=c | K=j, T=t) = Pr(C=c | K=j', T=t)
for any j, j' in c.
C3: masking probabilities are independent of theta
}
