% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_component_failure_probability.R
\name{md_series_component_failure_decorator}
\alias{md_series_component_failure_decorator}
\title{Decorate masked data (tbl_md) with component failure probabilities.}
\usage{
md_series_component_failure_decorator(md, Pk, var.t = "t", var.cand = "x")
}
\arguments{
\item{md}{masked data}

\item{Pk}{probability \code{Pr{K[i]=j|...}} for \code{j in {1,...,m}},}
}
\description{

}
\details{
We decorate masked data \code{md} with an estimate of the
probabilities, \code{Pr{K[i]=j|T[i]=t[i],C[i])}},
\code{Pr{K[i]=j|T[i]=t[i]}}, and \code{Pr{K[i]=j}} for \code{j in {1,...,m}}.
}
