% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_component_failure_probability.R
\name{md_series_component_failure_decorator_diff}
\alias{md_series_component_failure_decorator_diff}
\title{Decorate masked data \code{md} with a some measure of the
difference between component failure probabilities.}
\usage{
md_series_component_failure_decorator_diff(
  md,
  pk = "pk",
  diff = function(r) range(r)
)
}
\arguments{
\item{md}{masked data}

\item{pk}{prefix label of component failure probabilities}

\item{diff}{difference measure, defaults to range length}
}
\description{
Decorate masked data \code{md} with a some measure of the
difference between component failure probabilities.
}
