% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_component_failure_probability.R
\name{md_series_component_failure_probability_decorator}
\alias{md_series_component_failure_probability_decorator}
\title{Component failure probability decorator.}
\usage{
md_series_component_failure_probability_decorator(md, h)
}
\arguments{
\item{md}{masked data}

\item{h}{a vector of hazard functions for the component lifetimes in the
series system. \code{h[j]} should refer to the
hazard function for component indexed by \code{j}.}
}
\value{
\code{md} decorated with component probability failures.
}
\description{
We decorate masked data \code{md} with an estimate of the
probabilities, \code{Pr{K[i]=j|T[i]=t[i],C[i])}}.
}
\details{
\code{md} must have columns \code{x1},...,\code{xm} representing candidate
sets as boolean vectors, \code{s} or \code{t} respectively representing
right-censored series system lifetime, and if right-censoring, then there
must also be \code{delta} representing the right-censoring indicator.
}
