% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_series_data}
\alias{md_series_data}
\title{Generates masked data for a series system with the given node failure times
\code{t}, candidate set model \code{candidate_model}, and candidate
set sizes \code{w}.}
\usage{
md_series_data(t, w, candidate_model = md_candidate_m0)
}
\arguments{
\item{t}{matrix of node failure times}

\item{w}{Integer vector. For the ith observation, generate w_j candidates.}

\item{candidate_model}{Function that accepts masked data as an argument.
The candidate model, defaults to md_candidate_m0.
If set to NULL, then do not generate a candidate set.
md_mle_exp_series will treat such masked data as
a sample that includes every node as candidates.}
}
\value{
masked data, a data frame of n observations, (s,k,t1,...,tm,c1,...,cm)
where k, t, and c are covariates (or predictors) of s,k,t1,...,tm.
}
\description{
Generates masked data for a series system with the given node failure times
\code{t}, candidate set model \code{candidate_model}, and candidate
set sizes \code{w}.
}
