% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_decorators.R
\name{md_series_lifetime}
\alias{md_series_lifetime}
\title{Generates failure times and component cause of failure for a series
system with the given data frame of component lifetimes encoded by
the matrix columns prefixed with \code{t} in masked data frame \code{md}.}
\usage{
md_series_lifetime(md)
}
\arguments{
\item{md}{a data frame with the indicated component times-to-failure}
}
\value{
masked data decorated with series lifetime and component cause of failure
}
\description{
Generates failure times and component cause of failure for a series
system with the given data frame of component lifetimes encoded by
the matrix columns prefixed with \code{t} in masked data frame \code{md}.
}
