% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_decorators.R
\name{md_series_lifetime_right_censoring}
\alias{md_series_lifetime_right_censoring}
\title{Generates right-censored system failure times \code{s} and right-censoring
indicators \code{delta} for a series system with the given data frame of
component lifetimes encoded by the matrix columns prefixed with \code{t}}
\usage{
md_series_lifetime_right_censoring(md, tau)
}
\arguments{
\item{md}{a data frame with the indicated component times-to-failure}

\item{tau}{vector of right-censoring times}
}
\value{
masked data decorated with series lifetime and component cause of failure
}
\description{
Generates right-censored system failure times \code{s} and right-censoring
indicators \code{delta} for a series system with the given data frame of
component lifetimes encoded by the matrix columns prefixed with \code{t}
}
