% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_failure_estimator.R
\name{md_series_node_failure_decorator_m0}
\alias{md_series_node_failure_decorator_m0}
\title{Decorate masked data (tbl_md) with node failure probabilities.}
\usage{
md_series_node_failure_decorator_m0(md, h)
}
\arguments{
\item{md}{masked data}

\item{h}{hazard funct}
}
\description{
Under model m0, we do not know which node caused the failure,
assuming |C| > 1, but if we have an estimate of the node
hazard functions (e.g., an estimate of theta from masked data),
then we can estimate the probability of node failure.
}
\details{
We decorate masked data \code{md} with an estimate of the
probabilities, \code{f(k|s,c)} for \code{k=1,...,k=m}.
}
