% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_failure_estimator.R
\name{md_series_node_failure_m1}
\alias{md_series_node_failure_m1}
\title{A pdf generator of conditional node failure in a series
system acording to candidate mdoel m0, \code{f(k|s,c)}.}
\usage{
md_series_node_failure_m1(h)
}
\arguments{
\item{h}{vector of hazard functions for the nodes in series system.}
}
\description{
\code{f(k|s,c,alpha)} is proportional to \code{alpha h_k(s)/h(s) I(k in c) + (1-alpha) h_k(s)/h(s) I(k not in c)}.
}
