% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_system_failure.R
\name{md_series_system_failure_decorator_m0}
\alias{md_series_system_failure_decorator_m0}
\title{Decorate masked data (tbl_md) with node failure probabilities.}
\usage{
md_series_system_failure_decorator_m0(md, q)
}
\arguments{
\item{md}{masked data}

\item{q}{interval computer for s|c}
}
\description{
Under model m0, we do not know which node caused the failure,
(note: if |C|=1, under m0 we know precisely which node failed),
but if we have an estimate (or know) theta, then we may
construct f(k|s,c) and compute the node failure probabilities
in a masked data object \code{md}.
}
\details{
We decorate masked data \code{md} with an estimate of the
probabilities, \code{f(k|s,c)} for \code{k=1,...,k=m}
and return the result.
}
