% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_tools.R
\name{md_solver_gradient_ascent}
\alias{md_solver_gradient_ascent}
\title{Gradient ascent algorithm.}
\usage{
md_solver_gradient_ascent(
  theta0,
  grad,
  eps = 1e-05,
  max_iterations = 10000L,
  gamma = 0.1
)
}
\arguments{
\item{theta0}{initial guess of theta with \eqn{p} components}

\item{eps}{stopping condition}

\item{max_iterations}{maximum number of iterations}

\item{gamma}{step size}

\item{score}{score function of type \eqn{R^p -> R^p}}
}
\description{
Gradient ascent algorithm.
}
\section{Algorithm}{

The algorithm is straightforward. Details here.
}

\author{
Alex Towell
}
