% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{md_write_csv}
\alias{md_write_csv}
\title{Write masked data data frame (tibble) object to a CSV (comma separated file),
optionally writing associated meta-data to a JSON file. In particular, meta-data
in this case is defined as the attributes of the data frame object.}
\usage{
md_write_csv(md, filename, write.metadata = T)
}
\arguments{
\item{md}{a masked data frame}

\item{filename}{filename for csv}

\item{write.metadata}{write a separate}
}
\description{
Write masked data data frame (tibble) object to a CSV (comma separated file),
optionally writing associated meta-data to a JSON file. In particular, meta-data
in this case is defined as the attributes of the data frame object.
}
