% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{nparams.series}
\alias{nparams.series}
\title{Number of parameters for a general series system.}
\usage{
\method{nparams}{series}(x)
}
\arguments{
\item{x}{series object}
}
\value{
number of parameters
}
\description{
Number of parameters for a general series system.
}
