% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{num_nodes}
\alias{num_nodes}
\title{Method for obtaining the number of nodes in an object.}
\usage{
num_nodes(x)
}
\arguments{
\item{series}{The object to obtain the number of nodes of}
}
\description{
Method for obtaining the number of nodes in an object.
}
