% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_general_series.R
\name{numerical_mle}
\alias{numerical_mle}
\title{Maximum likelihood estimator given a log-likelihood function}
\usage{
numerical_mle(
  ll,
  theta0,
  info = NULL,
  reltol = sqrt(.Machine$double.eps),
  method = "BFGS",
  keep_obs = FALSE,
  max_iter = 250L,
  ...
)
}
\arguments{
\item{theta0}{initial value for the MLE}

\item{keep_obs}{logical, keep observations if TRUE}

\item{max_iter}{integer, maximum number of iterations}

\item{...}{pass additional arguments to \code{optim}}

\item{eps}{convergence tolerance (default is 1e-6)}
}
\description{
Maximum likelihood estimator given a log-likelihood function
}
