% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series_system.R
\name{param_weibull_series}
\alias{param_weibull_series}
\title{The cumulative distribution function for Weibull series}
\usage{
param_weibull_series(scales, shapes, g = function(t) t)
}
\arguments{
\item{scales}{scale parameters for Weibull component lifetimes}

\item{shapes}{shape parameters for Weibull component lifetimes}

\item{g}{function of system failure time \code{t}, some population
parameter of interest, such as the expected value or the variance.
defaults to expected value, \code{function(t) t}.}
}
\description{
The cumulative distribution function for Weibull series
}
