% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{params.exp_dist}
\alias{params.exp_dist}
\title{Method for obtaining the parameters of
a \code{exp_dist} distribution object.}
\usage{
\method{params}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The \code{exp_dist} object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of
a \code{exp_dist} distribution object.
}
