% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{params.normal}
\alias{params.normal}
\title{Method for obtaining the parameters of a \code{normal} object.}
\usage{
\method{params}{normal}(x, ...)
}
\arguments{
\item{x}{The object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of a \code{normal} object.
}
