% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{params}
\alias{params}
\title{Generic method for obtaining the parameters of
a parametric distribution.}
\usage{
params(x, ...)
}
\arguments{
\item{x}{The object to obtain the parameters of}
}
\description{
Generic method for obtaining the parameters of
a parametric distribution.
}
