% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{params.series}
\alias{params.series}
\title{Method for obtaining the parameters of
a \code{series} distribution object.}
\usage{
\method{params}{series}(x, ...)
}
\arguments{
\item{x}{The \code{series} object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of
a \code{series} distribution object.
}
