% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{pdf.exp_dist}
\alias{pdf.exp_dist}
\title{Method to obtain the pdf of an \code{exp_dist} object.}
\usage{
\method{pdf}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}
}
\description{
Note that since \code{exp_series} is also exponentially distributed,
this works for that too.
}
