% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{pdf.exp_series_dist}
\alias{pdf.exp_series_dist}
\title{Method to obtain the pdf of an \code{exp_series_dist} object.}
\usage{
\method{pdf}{exp_series_dist}(x, ...)
}
\arguments{
\item{x}{The object to obtain the pdf of}
}
\description{
Method to obtain the pdf of an \code{exp_series_dist} object.
}
