% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{pdf}
\alias{pdf}
\title{Generic method for obtaining the pdf function of
a random variable.}
\usage{
pdf(x, ...)
}
\arguments{
\item{x}{The object to obtain the hazard function of}
}
\description{
Generic method for obtaining the pdf function of
a random variable.
}
