% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system.R
\name{pexp_series_system}
\alias{pexp_series_system}
\title{Cumulative distribution function for exponential series.}
\usage{
pexp_series_system(t, rates, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{t}{Vector of series system lifetimes.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \verb{P[X <= x], otherwise, }P\link{X > x}`.}

\item{log.p}{Logical; if TRUE, return the log of the cdf.}
}
\value{
The cumulative probabilities evaluated at the specified lifetimes.
}
\description{
Cumulative distribution function for exponential series.
}
