% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{pgen_series}
\alias{pgen_series}
\title{cdf for general series}
\usage{
pgen_series(t, theta, nparams, R)
}
\arguments{
\item{t}{series system lifetime}

\item{theta}{parameter vector}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{R}{list of reliability functions}
}
\description{
cdf for general series
}
