% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{point.md_estimate}
\alias{point.md_estimate}
\title{Method to obtain the point estimate of
a masked data estimator, \code{md_estimate}.}
\usage{
\method{point}{md_estimate}(x, ...)
}
\arguments{
\item{x}{The \code{md_estimate} object to obtain the point estimate of}
}
\description{
Method to obtain the point estimate of
a masked data estimator, \code{md_estimate}.
}
