% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle.R
\name{point.md_mle}
\alias{point.md_mle}
\title{Method to obtain the point estimate of
a masked data estimator, \code{md_mle}.}
\usage{
\method{point}{md_mle}(x, ...)
}
\arguments{
\item{x}{The \code{md_mle} object to obtain the MLE point estimate of}

\item{...}{Additional arguments to pass.}
}
\description{
Method to obtain the point estimate of
a masked data estimator, \code{md_mle}.
}
