% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{point}
\alias{point}
\title{Generic method for obtaining the point estimate of
an estimator.}
\usage{
point(x, ...)
}
\arguments{
\item{x}{The object to obtain the point estimate of}
}
\description{
Generic method for obtaining the point estimate of
an estimator.
}
