% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tools.R
\name{print.tbl_md}
\alias{print.tbl_md}
\title{Print method for masked data (tbl_md).}
\usage{
\method{print}{tbl_md}(x, pprint = F, drop_latent = F, ...)
}
\arguments{
\item{x}{masked data to print}

\item{pprint}{Boolean, show candidates as a string column}

\item{drop_latent}{Boolean, drop the latent random variables}
}
\description{
Print method for masked data (tbl_md).
}
