% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{qexp_series}
\alias{qexp_series}
\title{Qunatile function for exponential series.}
\usage{
qexp_series(p, rates, lower.tail = T, log.p = F)
}
\arguments{
\item{p}{quantiles}

\item{rates}{rate parameters for exponential component lifetimes}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}, otherwise, \code{P[X > x]}.}

\item{log.p}{return the log of the survival function}
}
\description{
Qunatile function for exponential series.
}
