% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{qgeneral_series}
\alias{qgeneral_series}
\title{Quantile function (inverse of the cdf) for a general series system.
By definition, the quantile \code{p * 100\%} is the value \code{t} that
satisfies \code{F(t) - p = 0}. We solve for \code{t} using Newton's method.}
\usage{
qgeneral_series(p, theta, nparams, h, R, eps = 0.001, t0 = 1)
}
\arguments{
\item{p}{vector of probabilities.}

\item{theta}{parameter vector to evaluate h's and R's at}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{h}{list of hazard functions}

\item{R}{list of reliability functions}

\item{eps}{stopping condition, default is 1e-3}

\item{t0}{initial guess, default is 1}
}
\description{
Quantile function (inverse of the cdf) for a general series system.
By definition, the quantile \code{p * 100\%} is the value \code{t} that
satisfies \code{F(t) - p = 0}. We solve for \code{t} using Newton's method.
}
