% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcomp}
\alias{rcomp}
\title{Random generation for a component with custom hazard/survival}
\usage{
rcomp(n, haz, surv, theta)
}
\arguments{
\item{n}{number of samples to generate}

\item{haz}{hazard function h(t, theta)}

\item{surv}{survival function S(t, theta)}

\item{theta}{parameter vector passed to haz and surv}
}
\value{
vector of n random samples
}
\description{
Generates random samples using inverse transform sampling.
}
