% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_md_c1_c2_c3.R
\name{rdata.exp_series_md_c1_c2_c3}
\alias{rdata.exp_series_md_c1_c2_c3}
\title{Random data generation for \code{exp_series_md_c1_c2_c3} model.}
\usage{
\method{rdata}{exp_series_md_c1_c2_c3}(model, ...)
}
\arguments{
\item{model}{the likelihood model object}

\item{...}{additional arguments (passed to returned function)}
}
\value{
function that takes (theta, n, tau, p, observe, ...) and returns a
data frame with columns: t, omega, x1, x2, ..., xm
}
\description{
Returns a function that generates random masked data from the exponential
series system DGP at a given parameter value.
}
