% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_homogeneous_md_c1_c2_c3.R
\name{rdata.wei_series_homogeneous_md_c1_c2_c3}
\alias{rdata.wei_series_homogeneous_md_c1_c2_c3}
\title{Random data generation for \code{wei_series_homogeneous_md_c1_c2_c3} model.}
\usage{
\method{rdata}{wei_series_homogeneous_md_c1_c2_c3}(model, ...)
}
\arguments{
\item{model}{the likelihood model object}

\item{...}{additional arguments (passed to returned function)}
}
\value{
function that takes (theta, n, tau, p, ...) and returns a data frame
with columns: t, delta, x1, x2, ..., xm
}
\description{
Returns a function that generates random masked data from the homogeneous
Weibull series system DGP at a given parameter value.
}
