% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loglik}
\alias{score}
\alias{hess_loglik}
\alias{assumptions}
\alias{fim}
\alias{observed_info}
\alias{rdata}
\alias{fit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{likelihood.model}{\code{\link[likelihood.model]{assumptions}}, \code{\link[likelihood.model]{fim}}, \code{\link[likelihood.model]{hess_loglik}}, \code{\link[likelihood.model]{loglik}}, \code{\link[likelihood.model]{observed_info}}, \code{\link[likelihood.model]{rdata}}, \code{\link[likelihood.model]{score}}}
}}

