% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{rexp_series}
\alias{rexp_series}
\title{Survival function for exponential series.}
\usage{
rexp_series(n, rates)
}
\arguments{
\item{n}{sample size}

\item{rates}{rate parameters for exponential component lifetimes}
}
\description{
Survival function for exponential series.
}
