% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system.R
\name{rexp_series_system}
\alias{rexp_series_system}
\title{Survival function for exponential series.}
\usage{
rexp_series_system(n, rates, keep_latent = FALSE)
}
\arguments{
\item{n}{Integer, number of observations to generate.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}
}
\value{
A vector of random variates from the specified exponential series
distribution.
}
\description{
Survival function for exponential series.
}
