% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{rgeneral_series}
\alias{rgeneral_series}
\title{Sample from a general series system of \code{m} components whose hazard
and reliability functions are respectively given by \code{h} and \code{R}.}
\usage{
rgeneral_series(n, theta, h, R)
}
\arguments{
\item{n}{sample size}

\item{theta}{parameter vector to evaluate h's and R's at}

\item{h}{list of hazard functions}

\item{R}{list of reliability functions}
}
\description{
Sample from a general series system of \code{m} components whose hazard
and reliability functions are respectively given by \code{h} and \code{R}.
}
