% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked.data.R
\name{rmasked.data}
\alias{rmasked.data}
\title{rmask}
\usage{
rmasked.data(data, w = NULL, alpha = NULL)
}
\arguments{
\item{data}{a data frame, first column failure time, second column failed component, remaining columns component lifetimes.}

\item{w}{a vector of sizes for candidate sets}

\item{alpha}{a vector of probabilities}
}
\value{
masked data
}
\description{
Generate masked data from series system data
}
\examples{
n = 100
t = rseries.exp(n,c(1,2,3))
w = rep(2,n)
alpha = rep(.9,n)
data = rmasked.data(t,w,alpha)
}
