% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{rseries.exp.mle.cov.bootstrap}
\alias{rseries.exp.mle.cov.bootstrap}
\title{bootstrap of the mle's sampling distribution}
\usage{
rseries.exp.mle.cov.bootstrap(masked.data, r)
}
\arguments{
\item{masked.data}{sample of masked data}

\item{r}{replicates}
}
\value{
mle and bootstrap estimate of its covariance matrix
}
\description{
bootstrap of the mle's sampling distribution
}
