% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{rseries.exp}
\alias{rseries.exp}
\title{rseries.exp}
\usage{
rseries.exp(n, rate)
}
\arguments{
\item{n}{sample size (each row is an observation)}

\item{rate}{the j-th component has a failure rate rate_j}
}
\value{
matrix of n x length(rates) component lifetimes
}
\description{
rseries.exp
}
\examples{
# generate 10 samples (10 x 3 matrix of component lifetimes)
t = rseries.exp(n=10,rate=c(1,2,3))
}
