% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-lomax.R
\name{rseries.lomax}
\alias{rseries.lomax}
\title{rseries.lomax}
\usage{
rseries.lomax(n, lambda, kappa)
}
\arguments{
\item{n}{Numeric. Number of observations.}

\item{lambda}{Numberic vector. The j-th component has lambda=lambda_j}

\item{kappa}{Numberic vector. The j-th component has kappa=kappa_j}
}
\value{
matrix of n x length(lambda) component lifetimes
}
\description{
joint distribution of S,K,T1,...,Tm where
}
\details{
\preformatted{Tj ~ lomax(theta_j*)
S = min\{T1,...,Tm\}
K = argmin_k \{Tk : k=1,...,m\}
}
}
\examples{
# generate 10 samples (10 x 3 matrix of component lifetimes)
t = rseries.lomax(n=10,lambda=c(1,2,3),kappa=c(4,5,6))
}
