% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{sampler.exp_dist}
\alias{sampler.exp_dist}
\title{Method to sample from an \code{exp_dist} object.}
\usage{
\method{sampler}{exp_dist}(x, ...)
}
\arguments{
\item{x}{The object to sample from.}
}
\description{
Method to sample from an \code{exp_dist} object.
}
