% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{sampler.exp_series_dist}
\alias{sampler.exp_series_dist}
\title{Method to sample from an \code{exp_series_dist} object.}
\usage{
\method{sampler}{exp_series_dist}(x, ...)
}
\arguments{
\item{x}{The \code{exp_series_dist} object to sample from.}
}
\description{
Method to sample from an \code{exp_series_dist} object.
}
