% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{sampler.md_estimate}
\alias{sampler.md_estimate}
\title{Method to obtain the sampler for an \code{md_estimate} object.}
\usage{
\method{sampler}{md_estimate}(x, ...)
}
\arguments{
\item{x}{The \code{md_estimate} object to create a sampling procedure from}
}
\description{
Method to obtain the sampler for an \code{md_estimate} object.
}
