% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle.R
\name{sampler.md_mle}
\alias{sampler.md_mle}
\title{Method to obtain the sampler for an \code{md_mle} object.}
\usage{
\method{sampler}{md_mle}(x, ...)
}
\arguments{
\item{x}{The \code{md_mle} object to create a sampling procedure from}

\item{...}{Additional arguments to pass.}
}
\description{
Method to obtain the sampler for an \code{md_mle} object.
}
