% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{sampler.normal}
\alias{sampler.normal}
\title{Method for sampling from a \code{normal} object.}
\usage{
\method{sampler}{normal}(x, ...)
}
\arguments{
\item{x}{The object to sample from}
}
\description{
Method for sampling from a \code{normal} object.
}
