% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{sampler}
\alias{sampler}
\title{Generic method for sampling from distribution objects.}
\usage{
sampler(x, ...)
}
\arguments{
\item{x}{The object to sample from.}
}
\description{
Generic method for sampling from distribution objects.
}
