% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{sampler.series}
\alias{sampler.series}
\title{Sample from a general series system.}
\usage{
sampler.series(x, ...)
}
\arguments{
\item{...}{additional arguments to pass into \code{options}}

\item{n}{sample size}

\item{theta}{parameter vector, if NULL uses the values in \code{object}}

\item{object}{a \code{series} object}

\item{options}{list of options}
}
\description{
Sample from a general series system.
}
