% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{series.exp.info}
\alias{series.exp.info}
\title{information matrix of the rate parameter
of the series system with exponentially
distributed component lifetimes. this is
the expected info. if observed info is
desired, then just evaluate the hessian
of the log-likelihood function at the
mle.}
\usage{
series.exp.info(n, rate)
}
\arguments{
\item{n}{masked data sample size}

\item{rate}{true rate (or mle)}
}
\value{
expected info
}
\description{
information matrix of the rate parameter
of the series system with exponentially
distributed component lifetimes. this is
the expected info. if observed info is
desired, then just evaluate the hessian
of the log-likelihood function at the
mle.
}
