% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{series.exp.mle.cov}
\alias{series.exp.mle.cov}
\title{sampling distribution of the mle is
a multivariate normal with mean
given by the true rate parameterfor a series system
with exponentially distributed component
lifetimes, given a sample of masked data.}
\usage{
series.exp.mle.cov(n, rate)
}
\arguments{
\item{n}{sample size}

\item{rate}{true rate parameter (or mle)}
}
\value{
multivariate normal of the mle's sampling distribution
}
\description{
sampling distribution of the mle is
a multivariate normal with mean
given by the true rate parameterfor a series system
with exponentially distributed component
lifetimes, given a sample of masked data.
}
