% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-exp.R
\name{series.exp.mle}
\alias{series.exp.mle}
\title{maximum likelihood estimator of the
parameters of a series system with
components that have exponentially
distributed lifetimes given a
sample of masked data.}
\usage{
series.exp.mle(masked.data)
}
\arguments{
\item{masked.data}{masked data}
}
\value{
mle of rate parameter
}
\description{
maximum likelihood estimator of the
parameters of a series system with
components that have exponentially
distributed lifetimes given a
sample of masked data.
}
