% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series}
\alias{series}
\title{General series system object constructor.}
\usage{
series(hazs, nparams, survs = NULL)
}
\arguments{
\item{hazs}{list of hazard functions.}

\item{survs}{list of component survival (reliability) functions.
Defaults to NULL, in which case we generate the survival
functions from \code{hazs} (using cumulative hazard functions).}

\item{nparam}{integer vector, we have m components, each of which
may have a different number of parameters. the j-th
component has \code{nparams[j]} parameters.}
}
\value{
\code{series} system object
}
\description{
General series system object constructor.
}
