% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.system.estimation.masked.data.R
\docType{package}
\name{series.system.estimation.masked.data}
\alias{series.system.estimation.masked.data}
\title{series.system.masked.data: A package for estimating series system
component lifetime parameters from masked data}
\description{
The \code{series.system.estimation.masked.data} package provides a general
framework for estimating series system lifetime component parameters
from masked data.
}
