% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{series_data}
\alias{series_data}
\title{Generates system failure time and component failure for a series system with
the given matix of component failure times.}
\usage{
series_data(t)
}
\arguments{
\item{t}{matrix of node failure times}
}
\description{
Generates system failure time and component failure for a series system with
the given matix of component failure times.
}
