% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_md_c1_c2_c3.R
\name{series_md_c1_c2_c3}
\alias{series_md_c1_c2_c3}
\title{Likelihood model for series systems with masked component cause
of failure with candidate sets that satisfy conditions C1, C2, and C3,
desribed below. It is defined by hazard functions and
survival functions}
\usage{
series_md_c1_c2_c3(x)
}
\description{
This model satisfies the concept of a \code{likelihood_model} in the
\code{likelihood_model} package by providing the following methods:
}
\details{
(1) \code{loglik.exp_series_md_c1_c2_c3}
(2) \code{score.exp_series_md_c1_c2_c3}
(3) \code{hess_loglik.exp_system_series_md_c1_c2_c3}

These are useful for doing maximum likelihood estimation, hypothesis
testing (e.g., likelihood ratio test), estimation of asymptotic sampling
distribution given data from the DGP according to the specified model,
etc.

It is designed to work well with the \code{likelihood_model} R package. In
particular, it is intended to be used with the \code{likelihood_contr_model}
object, which is a \code{likelihood_model} object that allows likelihood
contributions to be added for whatever data model you have in mind.

In this likelihood model, masked component data approximately satisfies the
following conditions:

C1: \verb{Pr\{K[i] in C[i]) = 1}
C2: \verb{Pr\{C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])}
for any \verb{j, j' in c[i]}.
C3: masking probabilities are independent of \code{theta}

As a special case, this model also includes exact component cause of failure
data where the candidate set is a singleton.
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{hazard,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
\keyword{survival,}
