% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{surv.exp_series}
\alias{surv.exp_series}
\title{Survival function for exponential series.}
\usage{
surv.exp_series(t, rates, log.p = FALSE)
}
\arguments{
\item{t}{Vector of series system lifetimes.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}

\item{log.p}{Logical; if TRUE, return the log of the survival function.}
}
\value{
The survival function evaluated at the specified lifetimes.
}
\description{
Survival function for exponential series.
}
