% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{surv.series}
\alias{surv.series}
\title{Survival function for a general series system.}
\usage{
\method{surv}{series}(x)

\method{surv}{series}(x)
}
\arguments{
\item{t}{series system lifetime}

\item{theta}{parameter vector}

\item{object}{series object}

\item{log}{logical, log of survival function}
}
\value{
survival function
}
\description{
Survival function for a general series system.

Survival function for general series
}
