% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{survival_exp_series}
\alias{survival_exp_series}
\title{Survival function for exponential series.}
\usage{
survival_exp_series(t, rates, log.p = F)
}
\arguments{
\item{t}{series system lifetime}

\item{rates}{rate parameters for exponential component lifetimes}

\item{log.p}{return the log of the survival function}
}
\description{
Survival function for exponential series.
}
