% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{survival_general_series}
\alias{survival_general_series}
\title{Survival function for general series}
\usage{
survival_general_series(t, theta, nparams, R)
}
\arguments{
\item{t}{series system lifetime}

\item{theta}{parameter vector}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparams[j]} parameters.}

\item{R}{list of reliability functions}
}
\description{
Survival function for general series
}
\note{
convert this into a generator over (t,theta)
}
