% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series_system.R
\name{survival_weibull_series}
\alias{survival_weibull_series}
\title{Survival function for weibull series}
\usage{
survival_weibull_series(t, scales, shapes)
}
\arguments{
\item{t}{series system lifetime}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
Survival function for weibull series
}
