% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{vcov.exp_dist}
\alias{vcov.exp_dist}
\title{Method for obtaining the variance of a \code{exp_dist} object.}
\usage{
\method{vcov}{exp_dist}(object, ...)
}
\arguments{
\item{object}{The \code{exp_dist} object to obtain the variance of}
}
\description{
Method for obtaining the variance of a \code{exp_dist} object.
}
