% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{vcov.exp_series}
\alias{vcov.exp_series}
\title{Method for obtaining the variance-covariance of a \code{exp_series} object.}
\usage{
\method{vcov}{exp_series}(object, ...)
}
\arguments{
\item{object}{The \code{exp_series}The object to obtain the variance of}
}
\description{
Method for obtaining the variance-covariance of a \code{exp_series} object.
}
