% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{vcov.exp_series_dist}
\alias{vcov.exp_series_dist}
\title{Method for obtaining the variance-covariance of a \code{exp_series_dist} object.}
\usage{
\method{vcov}{exp_series_dist}(object, ...)
}
\arguments{
\item{object}{The \code{exp_series}The object to obtain the variance of}
}
\description{
Method for obtaining the variance-covariance of a \code{exp_series_dist} object.
}
