% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{vcov.md_estimate}
\alias{vcov.md_estimate}
\title{Compute the covariance matrix from the given masked data estimate.}
\usage{
\method{vcov}{md_estimate}(estimate)
}
\arguments{
\item{estimate}{the variance-covariance matrix of the estimator to obtain}
}
\description{
Sampling distribution of the MLE is a multivariate normal with mean
given by the true parameter value and, asymptotically, a covariance
given by the inverse of the Fisher information matrix.
}
