% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle.R
\name{vcov.md_mle}
\alias{vcov.md_mle}
\title{Compute the covariance matrix from the given masked data estimate.}
\usage{
\method{vcov}{md_mle}(object, ...)
}
\arguments{
\item{object}{The variance-covariance matrix of the estimator to obtain}

\item{...}{Additional arguments to pass.}
}
\description{
Sampling distribution of the MLE is a multivariate normal with mean
given by the true parameter value and, asymptotically, a covariance
given by the inverse of the Fisher information matrix.
}
