% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{vcov.normal}
\alias{vcov.normal}
\title{Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.}
\usage{
\method{vcov}{normal}(normal, ...)
}
\arguments{
\item{normal}{object to retrieve the variance-covariance matrix from}
}
\description{
Retrieve the variance-covariance matrix (or scalar)
of a \code{normal} object.
}
