% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_weibull_series_system.R
\name{Weibull series MLE}
\alias{Weibull series MLE}
\title{Maximum likelihood estimation functions for Weibull series systems
from masked data.}
\description{
Functions include the log-likelihood, score, and MLE functions.
}
\details{
Masked component data approximately satisfies the following conditions:
C1: Pr(K in C) = 1
C2: Pr(C=c | K=j, T=t) = Pr(C=c | K=j', T=t)
for any j, j' in c.
C3: masking probabilities are independent of theta
}
\seealso{
\code{\link{md_loglike_weibull_series_C1_C2_C3}},
\code{\link{md_score_weibull_series_C1_C2_C3}},
\code{\link{md_mle_weibull_series_C1_C2_C3}}
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
\keyword{weibull,}
