% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series_system.R
\name{Weibull series}
\alias{Weibull series}
\title{Weibull series}
\description{
This file contains functions related to the Weibull series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions for the Weibull series distribution.
}
\details{
For parameter estimation from masked data, see the file \code{md_weibull_series_mle.R}.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{series,}
\keyword{statistics}
\keyword{weibull,}
