% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{confint.md_estimate}
\alias{confint.md_estimate}
\title{Method to obtain the confidence intervals of the parameter values of a
masked data estimator, \code{md_estimate}.}
\usage{
\method{confint}{md_estimate}(estimate, parm = NULL, level = 0.95)
}
\arguments{
\item{level}{Confidence level, defaults to 0.95 (alpha=.05)}
}
\description{
Method to obtain the confidence intervals of the parameter values of a
masked data estimator, \code{md_estimate}.
}
