% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cum_haz}
\alias{cum_haz}
\title{cumulative hazard function for a component hazard function}
\usage{
cum_haz(haz)
}
\arguments{
\item{haz}{hazard function}
}
\description{
cumulative hazard function for a component hazard function
}
