% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_system_md_C1_C2_C3.R
\name{Exponential series likelihood model for masked component cause of
failure (candidate sets) satisfying conditions C1, C2, and C3 with exact
system failure time observations.}
\alias{Exponential series likelihood model for masked component cause of
failure (candidate sets) satisfying conditions C1, C2, and C3 with exact
system failure time observations.}
\alias{exp_series_system_md_C1_C2_C3}
\title{Likelihood model for exponential series systems with masked component cause
of failure with candidate sets that satisfy conditions C1, C2, and C3,
desribed below. It also includes right-censoring, but this is not the focus
of the model and normally we will disentangle the two and model them
as separate and independent contributions to the total likelihood model
under consideration.}
\usage{
exp_series_system_md_C1_C2_C3(rates = NULL)
}
\description{
Various functions are provided according to this model:
}
\details{
(1) \code{loglik.exp_series_system_md_C1_C2_C3}
(2) \code{score.exp_series_system_md_C1_C2_C3}
(3) \code{hess_loglik.exp_system_series_md_C1_C2_C3}
(4) \code{fim.exp_series_system_md_C1_C2_C3}

These are useful for doing maximum likelihood estimation, hypothesis
testing (e.g., likelihood ratio test), estimation of asymptotic sampling
distribution given data from the DGP according to the specified model,
etc.

It is designed to work well with the \code{algebraic.mle} package.

In this likelihood model, masked component data approximately satisfies the
following conditions:

C1: \verb{Pr\{K[i] in C[i]) = 1}
C2: \verb{Pr\{C[i]=c[i] | K[i]=j, T[i]=t[i]) = Pr(C[i]=c[i] | K[i]=j', T[i]=t[i])}
for any \verb{j, j' in c[i]}.
C3: masking probabilities are independent of \code{theta}

As a special case, this model also includes exact component cause of failure
data where the candidate set is a singleton.

In this likelihood model, we also assume that the failure time is exactly
known.
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{exponential,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
