% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_estimate.R
\name{fisher_info.md_estimate}
\alias{fisher_info.md_estimate}
\title{Method to obtain the fisher information matrix of an \code{md_estimate}.}
\usage{
\method{fisher_info}{md_estimate}(x, ...)
}
\arguments{
\item{x}{The \code{md_estimate} object to obtain the fisher information of}
}
\description{
Method to obtain the fisher information matrix of an \code{md_estimate}.
}
