% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dist.R
\name{hazard.exp_dist}
\alias{hazard.exp_dist}
\title{Method to obtain the hazard function of
an \code{exp_series} object.}
\usage{
\method{hazard}{exp_dist}(exp_dist, ...)
}
\arguments{
\item{exp_series}{The object to obtain the hazard function of}
}
\description{
Method to obtain the hazard function of
an \code{exp_series} object.
}
