% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_md_c1_c2_c3.R
\name{hess_loglik.exp_series_md_c1_c2_c3}
\alias{hess_loglik.exp_series_md_c1_c2_c3}
\title{Generates the observed information matrix (FIM) for an exponential series
system with respect to parameter \code{theta} for masked data with candidate
sets that approximately satisfy conditions C1, C2, and C3 with exact
failure times, or right-censoring (no component failure).}
\usage{
hess_loglik.exp_series_md_c1_c2_c3(model)
}
\arguments{
\item{model}{the exponential series system masked data model (competing risks)
for candidate set model that satisfies C1, C2, and C3}
}
\value{
hessian of the log-likelihood function \code{loglik.exp_series_md_c1_c2_c3}
that takes the following arguments:
\itemize{
\item \code{df}: masked data
\item \code{theta}: rate / scale parameters of component lifetime distributions
\item \code{sys.var}: system lifetime (optionally right-censored) column name,
defaults to \code{t}
\item \code{candset}: prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.
}
}
\description{
Generates the observed information matrix (FIM) for an exponential series
system with respect to parameter \code{theta} for masked data with candidate
sets that approximately satisfy conditions C1, C2, and C3 with exact
failure times, or right-censoring (no component failure).
}
