% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_models.R
\name{md_candidate_m0}
\alias{md_candidate_m0}
\title{Decorates masked data object md with candidate sets according to
candidate model m0.}
\usage{
md_candidate_m0(md)
}
\arguments{
\item{md}{masked data, data frame object with column 'k' for failed component
and column 'w' for corresponding candidate set size.}
}
\value{
masked data with candidate sets that model m0
}
\description{
Decorates masked data object md with candidate sets according to
candidate model m0.
}
