% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_models.R
\name{md_candidate_m1}
\alias{md_candidate_m1}
\title{Candidate model m1}
\usage{
md_candidate_m1(md, m)
}
\arguments{
\item{md}{masked data, a data frame object with column 'k' for failed component,
column 'w' for corresponding candidate set size, and column 'alpha' for
corresponding alpha probabilities}

\item{m}{Integer, number of components in the series system}
}
\value{
masked data
}
\description{
Decorates masked data object \code{md} with candidate sets according to relaxed
candidate model with \code{Pr{K[i] in C[i]} = alpha}.
}
