% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_candidate_models.R
\name{md_candidate_m1}
\alias{md_candidate_m1}
\title{Candidate model m1}
\usage{
md_candidate_m1(md, m)
}
\arguments{
\item{md}{masked data, a data frame object with column 'k' for failed component,
column 'w' for corresponding candidate set size, and column 'alpha' for
corresponding alpha probabilities}

\item{m}{Integer, number of nodes in the series system}
}
\value{
alpha-masked data with candidate sets that model m1
}
\description{
Decorates masked data object \code{md} with candidate sets according to candidate
model m1.
}
\details{
Specifically, the candidate sets are generated according to the alpha-masked
model, where with probability \code{alpha[i]}, \code{C[i,]} contains \code{k[i]}
and \code{w[i-1]} nodes randomly selected without replacement from
\code{{1,...,m} - { k[i]}} and with probability \code{1-alpha[i]},
\code{C[i,]} contains \code{w[i]} nodes randomly selected without
replacement from \code{{1,...,m} - { k[i] }}.
}
