% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_exp_series}
\alias{md_exp_series}
\title{Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.}
\usage{
md_exp_series(n, theta, w, candidate_model = md_candidate_m0, metadata = T)
}
\arguments{
\item{n}{Integer. The sample size (each row is an observation).}

\item{theta}{Numeric vector. The jth component has a failure rate \code{theta[j]}.}

\item{w}{Integer vector. For the ith observation, generate \code{w[j]} candidates.}

\item{candidate_model}{Function that accepts masked data as an argument.
The candidate model, defaults to \code{md_candidate_m0}.
If set to \code{NULL}, then do not generate a candidate set.
\code{md_mle_exp_series} will treat such masked data as
a sample that includes every node as candidates.}

\item{metadata}{Boolean. If \code{TRUE} writes meta-data for series system to
attributes of masked data.}
}
\value{
masked data, a data frame of n observations, \code{(s,k,t1,...,tm,c1,...,cm)}
where \code{k}, \code{t}, and \code{c} are covariates (or predictors) of
\code{s,k,t1,...,tm}.
}
\description{
Generates masked data for a series system with exponentially distributed
nodes and candidate sets according to candidate_model.
}
\examples{
md_exp_series(n=10,theta=c(1,2,3),w=rep(2,10))
}
