% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_tools.R
\name{md_fisher_scoring}
\alias{md_fisher_scoring}
\title{Fisher scoring algorithm.}
\usage{
md_fisher_scoring(theta0, info, score, eps = 1e-05, max_iterations = 10000L)
}
\arguments{
\item{theta0}{initial guess of theta with \eqn{p} components}

\item{info}{information matrix function of type \eqn{R^p -> R^{p \times q}}}

\item{score}{score function of type \eqn{R^p -> R^p}}

\item{eps}{stopping condition}

\item{max_iterations}{maximum number of iterations}
}
\description{
Fisher scoring algorithm.
}
\section{Algorithm}{

The algorithm is straightforward. Details here.
}

