% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series_system.R
\name{md_loglike_exp_series_C1_C2_C3}
\alias{md_loglike_exp_series_C1_C2_C3}
\title{Generates a log-likelihood for an exponential series system with respect to
parameter \code{theta} for masked data with candidate sets that satisfy conditions
C1, C2, and C3.}
\usage{
md_loglike_exp_series_C1_C2_C3(md, setvar = "x", sysvar = "t")
}
\arguments{
\item{md}{masked data}

\item{setvar}{prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.}

\item{sysvar}{system lifetime (optionally right-censored) column name}
}
\description{
Generates a log-likelihood for an exponential series system with respect to
parameter \code{theta} for masked data with candidate sets that satisfy conditions
C1, C2, and C3.
}
