% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series.R
\name{md_loglike_weibull_series_C1_C2_C3}
\alias{md_loglike_weibull_series_C1_C2_C3}
\title{Log-likelihood generator for Weibull series system on masked data,
where the masked data is in the form of right-censored system lifetimes
and masked component cause of failure.}
\usage{
md_loglike_weibull_series_C1_C2_C3(md)
}
\arguments{
\item{md}{masked data}
}
\value{
a log-likelihood function with respect to theta given md
}
\description{
Masked component data approximately satisfies the following conditions:
C1: Pr(K in C) = 1
C2: Pr(C=c | K=j, T=t) = Pr(C=c | K=j', T=t)
for any j, j' in c.
C3: masking probabilities are independent of theta
}
