% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_weibull_series_system.R
\name{md_loglike_weibull_series_C1_C2_C3}
\alias{md_loglike_weibull_series_C1_C2_C3}
\title{Log-likelihood generator for Weibull series system on masked data,
where the masked data is in the form of right-censored system lifetimes
and masked component cause of failure.}
\usage{
md_loglike_weibull_series_C1_C2_C3(md)
}
\arguments{
\item{md}{masked data}
}
\value{
a log-likelihood function with respect to theta given md
}
\description{
Log-likelihood generator for Weibull series system on masked data,
where the masked data is in the form of right-censored system lifetimes
and masked component cause of failure.
}
