% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax_series.R
\name{md_lomax_series}
\alias{md_lomax_series}
\title{joint distribution of S,K,T1,...,Tm where}
\usage{
md_lomax_series(n, lambda, kappa, w, candidate_model = md_candidate_m0)
}
\arguments{
\item{n}{Integer. Number of observations.}

\item{lambda}{Numeric vector.}

\item{kappa}{Numeric vector. The jth node is parameterized by theta_j := (lambda_j,kappa_j).}

\item{w}{Integer vector. For the ith observation, generate w_j candidates.}

\item{candidate_model}{the candidate model, defaults to md_candidate_m0}
}
\value{
masked data
}
\description{
\preformatted{tj ~ lomax(theta_j*)
s = min\{T1,...,Tm\}
k = argmin_k \{Tk : k=1,...,m\}
}
}
\examples{
md <- md_lomax_series(
    n=10,
    lambda=c(1,2,3),
    kappa=c(4,5,6),
    w=rep(2,10))
}
