% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{md_mle_exp_series_C1_C2_C3}
\alias{md_mle_exp_series_C1_C2_C3}
\title{Maximum likelihood estimator for exponential series
in which each component (or competing risk) is
parameterized by a single rate parameter.}
\usage{
md_mle_exp_series_C1_C2_C3(md, theta0, eps = 1e-06)
}
\arguments{
\item{md}{right-censored, masked data}

\item{theta0}{initial value for the MLE}

\item{eps}{convergence tolerance
(default is \code{1e-6})}
}
\description{
The sample (at least approximately) satisfy
competing risks conditions C1, C2, and C3 and
is failure time data may be right censored.
}
