% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series.R
\name{md_mle_exp_series_C1_C2_C3}
\alias{md_mle_exp_series_C1_C2_C3}
\title{Maximum likelihood estimator for exponential series
in which each component (or competing risk) is
parameterized by a single rate parameter.}
\usage{
md_mle_exp_series_C1_C2_C3(md, theta0, control = list(), ...)
}
\arguments{
\item{md}{right-censored masked data}

\item{theta0}{initial value for the MLE}

\item{control}{list of control parameters
\itemize{
\item eps convergence tolerance (default is 1e-6)
\item keep_obs logical, keep observations if TRUE
\item maxit integer, maximum number of iterations
\item compute_converged logical, whether to compute convergence info
}}

\item{...}{pass additional arguments to \code{control}}
}
\value{
MLE of type \code{md_mle_exp_series_C1_C2_C3}
}
\description{
The sample (at least approximately) satisfy
competing risks conditions C1, C2, and C3 and
is failure time data may be right censored.
}
