% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_tools.R
\name{md_num_comp}
\alias{md_num_comp}
\title{Method for obtaining the number of nodes in an object.}
\usage{
md_num_comp(series)
}
\arguments{
\item{series}{The object to obtain the number of nodes of}
}
\description{
Method for obtaining the number of nodes in an object.
}
