% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_lifetime.R
\name{md_series_lifetime_right_censoring}
\alias{md_series_lifetime_right_censoring}
\title{Generates right-censored system failure times and right-censoring
indicators for a series system with the given data frame of
component lifetimes.}
\usage{
md_series_lifetime_right_censoring(
  md,
  tau = Inf,
  compvar = "t",
  sysvar = NULL,
  delta = "delta"
)
}
\arguments{
\item{md}{a data frame with the indicated component times-to-failure}

\item{tau}{vector of right-censoring times, defaults to
\code{Inf} (no right censoring)}

\item{compvar}{component lifetimes}

\item{sysvar}{right-censored system lifetime}

\item{delta}{right-censoring indicator, defaults to "delta"}
}
\value{
masked data with right-censoring related columns
}
\description{
Generates right-censored system failure times and right-censoring
indicators for a series system with the given data frame of
component lifetimes.
}
