% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{mean.exp_series}
\alias{mean.exp_series}
\title{Mean function for exponential series.}
\usage{
\method{mean}{exp_series}(rates)
}
\arguments{
\item{rates}{Vector of rate parameters for exponential component lifetimes.}
}
\value{
The mean of the exponential series distribution.
}
\description{
Mean function for exponential series.
}
