% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qcomp}
\alias{qcomp}
\title{Quantile function for a component with custom survival function}
\usage{
qcomp(
  p,
  surv,
  theta,
  t_lower = .Machine$double.eps,
  t_upper = .Machine$double.xmax^0.5,
  ...
)
}
\arguments{
\item{p}{probability (quantile level), must be in (0, 1)}

\item{surv}{survival function S(t, theta, ...)}

\item{theta}{parameter vector passed to surv}

\item{t_lower}{lower bound for search interval}

\item{t_upper}{upper bound for search interval (sqrt to avoid overflow)}

\item{...}{additional arguments passed to surv}
}
\value{
time t such that S(t) = p
}
\description{
Finds the time t such that S(t) = p using root finding.
The survival function S(t) is assumed to be monotonically decreasing
from S(0) = 1 to S(inf) = 0.
}
\examples{
# Exponential survival function
surv_exp <- function(t, theta) exp(-theta * t)

# Median lifetime (50th percentile) for rate = 2
qcomp(0.5, surv = surv_exp, theta = 2.0)
}
