% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{qexp_series}
\alias{qexp_series}
\title{Quantile function for exponential series.}
\usage{
qexp_series(p, rates, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}

\item{lower.tail}{Logical, if TRUE (default), probabilities are P(X <= x),
otherwise, P(X > x).}

\item{log.p}{Logical, if TRUE, vector of probabilities \code{p} are given as \code{log(p)}.}
}
\value{
Quantiles corresponding to the given probabilities \code{p}.
}
\description{
Quantile function for exponential series.
}
