% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcomp}
\alias{rcomp}
\title{Random generation for a component with custom survival function}
\usage{
rcomp(n, surv, theta)
}
\arguments{
\item{n}{number of samples to generate}

\item{surv}{survival function S(t, theta)}

\item{theta}{parameter vector passed to surv}
}
\value{
vector of n random samples
}
\description{
Generates random samples using inverse transform sampling.
}
\examples{
# Exponential survival function
surv_exp <- function(t, theta) exp(-theta * t)

# Generate 10 random samples with rate = 2
set.seed(123)
rcomp(10, surv = surv_exp, theta = 2.0)
}
