% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{rexp_series}
\alias{rexp_series}
\title{Survival function for exponential series.}
\usage{
rexp_series(n, rates)
}
\arguments{
\item{n}{Integer, number of observations to generate.}

\item{rates}{Vector of rate parameters for exponential component lifetimes.}
}
\value{
A vector of random variates from the specified exponential series distribution.
}
\description{
Survival function for exponential series.
}
