% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_series.R
\name{rgeneral_series}
\alias{rgeneral_series}
\title{Sample from a general series system of \code{m} components whose hazard
and reliability functions are respectively given by \code{h} and \code{R}.}
\usage{
rgeneral_series(n, theta, nparams, h, R = NULL)
}
\arguments{
\item{n}{sample size}

\item{theta}{parameter vector}

\item{nparams}{integer vector, we have m components, each of which may have a
different number of parameters. the j-th component has
\code{nparmas[j]} parameters.}

\item{h}{list of hazard functions}

\item{R}{list of reliability functions, defaults to NULL (results in a numerical approximation from h)}
}
\description{
Sample from a general series system of \code{m} components whose hazard
and reliability functions are respectively given by \code{h} and \code{R}.
}
