% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_series_system.R
\name{rweibull_series}
\alias{rweibull_series}
\title{Sampler for weibull series.}
\usage{
rweibull_series(n, scales, shapes)
}
\arguments{
\item{n}{sample size}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
Sampler for weibull series.
}
