% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_md_c1_c2_c3.R
\name{score.exp_series_md_c1_c2_c3}
\alias{score.exp_series_md_c1_c2_c3}
\title{Score method for \code{exp_series_md_c1_c2_c3} model.}
\usage{
\method{score}{exp_series_md_c1_c2_c3}(model, ...)
}
\arguments{
\item{model}{the likelihood model object}

\item{...}{additional arguments (passed to returned function)}
}
\value{
score function that takes the following arguments:
\itemize{
\item \code{df}: masked data frame
\item \code{par}: rate parameters of exponential component lifetime distributions
\item \code{lifetime}: system lifetime column name (default from model)
\item \code{lifetime_upper}: interval upper bound column name (default from model)
\item \code{omega}: observation type column name (default from model)
\item \code{candset}: prefix of Boolean matrix encoding candidate sets
}
}
\description{
Returns a score (gradient) function for an exponential series system with
respect to parameter \code{theta} for masked component failure with candidate
sets that satisfy conditions C1, C2, and C3.
}
\details{
All four observation types (exact, right, left, interval) have closed-form
score contributions.
}
